% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gee.R
\name{addEEImage}
\alias{addEEImage}
\title{Add an Earth Engine image to a Leaflet map}
\usage{
addEEImage(
  map,
  image,
  bands = NULL,
  min = 0,
  max = 1,
  palette = NULL,
  group = NULL,
  ...
)
}
\arguments{
\item{map}{A Leaflet widget.}

\item{image}{Earth Engine image.}

\item{bands}{One or three bands.}

\item{min, max}{Visualization limits.}

\item{palette}{Optional palette.}

\item{group}{Layer group.}

\item{...}{Passed to \code{\link[leaflet:addTiles]{leaflet::addTiles()}}.}
}
\value{
A Leaflet widget.
}
\description{
Add an Earth Engine image to a Leaflet map
}
