% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rasterProducts.R
\name{aggregateLevel3}
\alias{aggregateLevel3}
\title{Aggregate a GEDI Level 3 raster}
\usage{
aggregateLevel3(x, fact, fun = mean, ...)
}
\arguments{
\item{x}{A raster path or \code{SpatRaster}.}

\item{fact}{Aggregation factor.}

\item{fun}{Aggregation function.}

\item{...}{Passed to \code{\link[terra:aggregate]{terra::aggregate()}}.}
}
\value{
A \code{\link[terra:SpatRaster-class]{terra::SpatRaster}}.
}
\description{
Aggregate a GEDI Level 3 raster
}
