% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gee.R
\name{aspect}
\alias{aspect}
\title{Earth Engine terrain aspect}
\usage{
aspect(x)
}
\arguments{
\item{x}{Earth Engine elevation image.}
}
\value{
An Earth Engine image.
}
\description{
Earth Engine terrain aspect
}
