% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gee.R
\name{build_ee_forest}
\alias{build_ee_forest}
\title{Train a regression forest in Earth Engine}
\usage{
build_ee_forest(training, response, predictors, trees = 500L, seed = 1L)
}
\arguments{
\item{training}{Earth Engine \code{FeatureCollection} containing predictors and
the response.}

\item{response}{Response property name, such as \code{"agbd"} or \code{"rh98"}.}

\item{predictors}{Predictor property names.}

\item{trees}{Number of trees.}

\item{seed}{Random seed.}
}
\value{
A trained Earth Engine classifier configured for regression.
}
\description{
Train a regression forest in Earth Engine
}
