% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clipLevel1Bdt.R
\name{cliplevel1Bdt}
\alias{cliplevel1Bdt}
\title{Clip GEDI level1Bdt data}
\usage{
cliplevel1Bdt(level1Bdt, extent)
}
\arguments{
\item{level1Bdt}{level1Bdt; S4 object of class "gedi.level1b.dt"}

\item{extent}{Extent object of a Spatial object .}
}
\value{
Returns An object of class "gedi.level1b.dt"; subset of GEDI Level1B data
}
\description{
Clip GEDI level1Bdt data within a given bounding coordinates
}
\examples{
#
## GEDI level1B file path
# level1_filepath = system.file("extdata", "lvis_level1_clip.h5", package="rLVIS")
#
## Reading LVIS level 2 file
# level1_waveform = readLevel1b(level1Bfilepath)
#
## Rectangle area for cliping
# xmin = -116.4683
# xmax = -116.3177
# ymin = 46.75208
# ymax = 46.84229
#
## creating an exent object
# ext<-extent(c(xmin=xmin, xmax=xmax, ymin=ymin, ymax=ymax))
#
# clipped_level1Bdt = clipLevel1(level1Bdt, extent=ext)
#
# library(leaflet)
# leaflet() \%>\%
# addCircleMarkers(clipped_level1Bdt@dt$longitude_bin0,
#                   clipped_level1Bdt@dt$latitude_bin0,
#                   radius = 1,
#                   opacity = 1,
#                   color = "red")  \%>\%
# addScaleBar(options = list(imperial = FALSE)) \%>\%
# addProviderTiles(providers$Esri.WorldImagery)
}
