% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clipLevel1Bdt.R
\name{cliplevel1BdtGeometry}
\alias{cliplevel1BdtGeometry}
\title{Clip GEDI level1Bdt data by geometry}
\usage{
cliplevel1BdtGeometry(level1Bdt, polygon_spdf)
}
\arguments{
\item{level1Bdt}{level1Bdt; S4 object of class "gedi.level1b.dt"}

\item{polygon_spdf}{SpatialDataFrame. A polygon dataset for clipping the waveform}
}
\value{
Returns An object of class "gedi.level1b.dt"; subset of GEDI Level1B data
}
\description{
Clip GEDI level1Bdt data within a given geometry area
}
\examples{
#
# #' GEDI level1B file path
# level1Bfilepath = system.file("extdata", "lvis_level1_clip.h5", package="rGEDI")
#
# #' Reading GEDI level1B file
# level1b = readLevel1b(level1Bfilepath)
#
# #' Creating GEDI level1Bdt object
# level1Bdt = level1Bdt(level1b)
#
# # Polgons file path
# polygon_filepath <- system.file("extdata", "clip_polygon.shp", package="rGEDI")
#
# # Reading GEDI level2B file
# polygon_spdf<-raster::shapefile(polygons_filepath)
#
# clipped_level1Bdt = clipLevel1Geometry(level1Bdt, polygon_spdf)
#
# library(leaflet)
# leaflet() \%>\%
#  addCircleMarkers(clipped_level1Bdt@dt$longitude_bin0,
#                   clipped_level1Bdt@dt$latitude_bin0,
#                   radius = 1,
#                   opacity = 1,
#                   color = "red")  \%>\%
#  addScaleBar(options = list(imperial = FALSE)) \%>\%
#  addPolygons(data=polygon_spdf,weight=1,col = 'white',
#              opacity = 1, fillOpacity = 0) \%>\%
#  addProviderTiles(providers$Esri.WorldImagery)
}
