% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clipLevel1B.R
\name{clipLevel1BGeometry}
\alias{clipLevel1BGeometry}
\title{Clip GEDI Level1B data by geometry}
\usage{
clipLevel1BGeometry(level1b, polygon, output = "", split_by = NULL)
}
\arguments{
\item{level1b}{A \code{\link[=gedi.level1b-class]{gedi.level1b}} object (output of \code{\link[=readLevel1B]{readLevel1B()}} function).
An S4 object of class "gedi.level1b".}

\item{polygon}{Polygon or Multipolygon. An object opened with \code{sf::st_read},}

\item{output}{Optional character path where to save the new
\code{\link[hdf5r:H5File-class]{hdf5r::H5File}}. The default stores a temporary file only.}

\item{split_by}{Polygon id. If defined, GEDI data will be clipped by each polygon using
the attribute specified by \code{split_by} from the attribute table.}
}
\value{
Returns a list of S4 object of class \code{\link[=gedi.level1b-class]{gedi.level1b}} containing clipped
GEDI Level1B data.
}
\description{
This function clips GEDI Level1B (geolocated waveforms) data within a
given bounding geometry
}
\examples{
\dontrun{
outdir <- tempdir()

# Specifying the path to GEDI level1B data (zip file)
level1B_fp_zip <- system.file("extdata",
  "GEDI01_B_2019108080338_O01964_T05337_02_003_01_sub.zip",
  package = "rGEDI"
)

# Unzipping GEDI level1B data
level1Bpath <- unzip(level1B_fp_zip, exdir = outdir)

# Reading GEDI level1B data (h5 file)
level1b <- readLevel1B(level1Bpath = level1Bpath)

# Specifying the path to shapefile
polygon_filepath <- system.file("extdata", "stands_cerrado.shp", package = "rGEDI")

# Reading shapefile as sf object
library(sf)
polygon <- sf::st_read(polygon_filepath)

# Spepecifing output file and path
output <- file.path(outdir, "GEDI01_B_2019108080338_O01964_T05337_02_003_01_clip")

# clipping GEDI Level1B data by extent boundary box
level1b_clip <- clipLevel1BGeometry(level1b,
  polygon = polygon,
  output = output,
  split_by = "id"
)

close(level1b)
lapply(level1b_clip, close)
}
}
