% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clipLevel2A.R
\name{cliplevel2ah5}
\alias{cliplevel2ah5}
\title{Clip GEDI level2a data}
\usage{
cliplevel2ah5(level2a, xleft, xright, ybottom, ytop, output = "")
}
\arguments{
\item{level2a}{h5file; S4 object of class H5File}

\item{xleft}{numeric. left x coordinates of rectangles.}

\item{xright}{numeric. right x coordinates of rectangles.}

\item{ybottom}{numeric. bottom y coordinates of rectangles.}

\item{ytop}{numeric. top y coordinates of rectangles.}

\item{output}{optional character path where to save the new h5file. Default "" (temporary file).}
}
\value{
Returns An object of class H5File; subset of LVIS Level2 data
}
\description{
Clip GEDI Level1 data within a given bounding coordinates
}
\examples{

#' LVIS level 2A file path
#level2apath = system.file("extdata", "lvis_level1_clip.h5", package="rLVIS")

# Rectangle
#xleft = 81
#xright = 83
#ybottom = 2
#ytop = 4

#' Reading LVIS level 2 file
#level1_waveform = readlevel2a(level2apath)

#output = tempfile(fileext=".h5")

#clipped_waveform = cliplevel2a(level1_waveform, output, xleft, xright, ybottom, ytop)

}
\author{
Caio Hamamura
}
