% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clipLevel2A.R
\name{cliplevel2ah5Geometry}
\alias{cliplevel2ah5Geometry}
\title{Clip LVIS Level1 data by geometry}
\usage{
cliplevel2ah5Geometry(level2a, polygon_spdf, output = "")
}
\arguments{
\item{level2a}{h5file; S4 object of class H5File}

\item{polygon_spdf}{SpatialDataFrame. A polygon dataset for clipping the waveform}

\item{output}{optional character path where to save the new h5file. Default "" (temporary file).}
}
\value{
Returns An object of class H5File; subset of LVIS Level1 data
}
\description{
Clip LVIS Level1 data within a given bounding coordinates
}
\examples{

#' LVIS level 2 file path
#level1_filepath = system.file("extdata", "biomas.zip", package="rGEDI")

#' Reading LVIS level 2 file
#level1_waveform = readlevel2a(level1_filepath)

# Polgons file path
#polygons_filepath <- system.file("extdata", "LVIS_Mondah_clip_polygon.shp", package="rLVIS")

# Reading LVIS level 2 file
#polygon_spdf<-raster::shapefile(polygons_filepath)

#output = tempfile(fileext="h5")

#clipped_waveform = clipLevel1Geometry(level1_waveform, output, polygon_spdf)

}
\author{
Caio Hamamura
}
