% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clipLevel2BPAVDProfile.R
\name{clipLevel2BPAVDProfile}
\alias{clipLevel2BPAVDProfile}
\title{Clip GEDI Plant Area Volume Density profile by Coordinates}
\usage{
clipLevel2BPAVDProfile(level2BPAVDProfile, xmin, xmax, ymin, ymax)
}
\arguments{
\item{level2BPAVDProfile}{A GEDI Level2B object (output of \code{\link[=getLevel2BPAVDProfile]{getLevel2BPAVDProfile()}} function).
An S4 object of class "data.table".}

\item{xmin}{Numeric. West longitude (x) coordinate of the bounding rectangle, in decimal degrees.}

\item{xmax}{Numeric. East longitude (x) coordinate of the bounding rectangle, in decimal degrees.}

\item{ymin}{Numeric. South latitude (y) coordinate of the bounding rectangle, in decimal degrees.}

\item{ymax}{Numeric. North latitude (y) coordinate of the bounding rectangle, in decimal degrees.}
}
\value{
Returns an S4 object of class \link[data.table:data.table]{data.table::data.table}
containing the Plant Area Volume Density profile data.
}
\description{
This function clips GEDI level2B derived
Plant Area Volume Density profile within a given bounding coordinates
}
\examples{
# specify the path to GEDI level2B data (zip file)
outdir <- tempdir()
level2B_fp_zip <- system.file("extdata",
  "GEDI02_B_2019108080338_O01964_T05337_02_001_01_sub.zip",
  package = "rGEDI"
)

# Unzipping GEDI level2A data
level2Bpath <- unzip(level2B_fp_zip, exdir = outdir)

# Reading GEDI level2B data (h5 file)
level2b <- readLevel2B(level2Bpath = level2Bpath)

# Extracting GEDI Plant Area Volume Density profile
level2BPAVDProfile <- getLevel2BPAVDProfile(level2b)

# Bounding rectangle coordinates
xmin <- -44.15036
xmax <- -44.10066
ymin <- -13.75831
ymax <- -13.71244

# Clipping GEDI Plant Area Volume Density profile by boundary box extent
level2BPAVDProfile_clip <- clipLevel2BPAVDProfile(level2BPAVDProfile, xmin, xmax, ymin, ymax)

close(level2b)
}
\seealso{
\url{https://www.earthdata.nasa.gov/data/catalog/lpcloud-gedi02-b-002}
}
