% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/level4A.R
\name{clipLevel4A}
\alias{clipLevel4A}
\title{Clip GEDI Level 4A footprints by an extent}
\usage{
clipLevel4A(level4A, xmin, xmax, ymin, ymax)
}
\arguments{
\item{level4A}{A table returned by \code{\link[=getLevel4A]{getLevel4A()}}.}

\item{xmin, xmax, ymin, ymax}{Bounding coordinates in longitude/latitude.}
}
\value{
A \link[data.table:data.table]{data.table::data.table}.
}
\description{
Clip GEDI Level 4A footprints by an extent
}
