% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/level4A.R
\name{clipLevel4AGeometry}
\alias{clipLevel4AGeometry}
\title{Clip GEDI Level 4A footprints by polygons}
\usage{
clipLevel4AGeometry(level4A, polygon, split_by = NULL)
}
\arguments{
\item{level4A}{A table returned by \code{\link[=getLevel4A]{getLevel4A()}}.}

\item{polygon}{An \code{sf}, \code{sfc}, or \code{SpatVector} polygon object.}

\item{split_by}{Optional polygon attribute copied to \code{poly_id}.}
}
\value{
A \link[data.table:data.table]{data.table::data.table}.
}
\description{
Clip GEDI Level 4A footprints by polygons
}
