% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rasterProducts.R
\name{clipLevel4B}
\alias{clipLevel4B}
\title{Clip a GEDI Level 4B raster}
\usage{
clipLevel4B(x, y, filename = "", overwrite = FALSE)
}
\arguments{
\item{x}{A raster path or \code{SpatRaster}.}

\item{y}{An extent \code{c(xmin, xmax, ymin, ymax)} or polygon object.}

\item{filename}{Optional output filename.}

\item{overwrite}{Logical.}
}
\value{
A \code{\link[terra:SpatRaster-class]{terra::SpatRaster}}.
}
\description{
Clip a GEDI Level 4B raster
}
