% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clipLevel1BGeo.R
\name{clipx}
\alias{clipx}
\title{Clip GEDI x data}
\usage{
clipx(x, xleft, xright, ybottom, ytop)
}
\arguments{
\item{x}{x; S4 object of class "gedi.level1b.dt"}

\item{extent}{Extent object of a Spatial object .}
}
\value{
Returns An object of class "gedi.level1b.dt"; subset of GEDI Level1B data
}
\description{
Clip GEDI x data within a given bounding coordinates
}
\examples{

#' GEDI level1B file path
level1_filepath = system.file("extdata", "lvis_level1_clip.h5", package="rLVIS")

#' Reading LVIS level 2 file
level1_waveform = readLevel1b(level1Bfilepath)

# Rectangle area for cliping
xmin = -116.4683
xmax = -116.3177
ymin = 46.75208
ymax = 46.84229

# creating an exent object
ext<-extent(c(xmin=xmin, xmax=xmax, ymin=ymin, ymax=ymax))

clipped_x = clipLevel1(x, extent=ext)

library(leaflet)
leaflet() \%>\%
 addCircleMarkers(clipped_x@dt$longitude_bin0,
                  clipped_x@dt$latitude_bin0,
                  radius = 1,
                  opacity = 1,
                  color = "red")  \%>\%
 addScaleBar(options = list(imperial = FALSE)) \%>\%
 addProviderTiles(providers$Esri.WorldImagery)
}
