<<<<<<< HEAD
% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clipLevel1BGeo.R
\name{clipxGeometry}
\alias{clipxGeometry}
\title{Clip Level1BGeo data by geometry}
\usage{
clipLevel1BGeo(level1BGeo, polygon_spdf, output)
}
\arguments{
\item{level1BGeo}{A GEDI Level1b object (output of \code{\link[rGEDI:readLevel1B]{readLevel1B}} function). A S4 object of class "gedi.level1b".}

\item{polygon_spdf}{Polygon. An object of class \code{\link[sp]{SpatialPolygonsDataFrame-class}},
which can be loaded as an ESRI shapefile using \code{\link[rgdal:readOGR]{readOGR}} function in the \emph{rgdal} package.}

\item{split_by}{Polygon id. If defined, GEDI data will be clipped by each polygon using the polygon id from table of attribute defined by the user}
}
\value{
A S4 object of class \code{\link[data.table:data.table]{data.table-class}}.
}
\description{
This function clips GEDI level1B extracted geolocation (level1BGeo)
data within given geometry
}
\examples{

# specify the path to GEDI Level 1B data
level1bpath <- system.file("extdata", "GEDIexample_level01B.h5", package="rGEDI")

# Reading GEDI level1B data
level1b <- readLevel1B(level1bpath)

# Get GEDI level1B geolocations
level1BGeo<-getLevel1BGeo(level1b)

# specify the path to shapefile
polygon_filepath <- system.file("extdata", "clip_polygon.shp", package="rGEDI")

# Reading shapefile as SpatialPolygonsDataFrame object
library(rgdal)
polygon_spdf<-readOGR(polygons_filepath)

level1bGeo_clip = clipLevel1BGeometry(level1bGeo, polygon_spdf, split_by="id")

library(leaflet)
leaflet() \%>\%
 addCircleMarkers(level1b_clip@dt$longitude_bin0,
                  level1b_clip@dt$latitude_bin0,
                  radius = 1,
                  opacity = 1,
                  color = "red")  \%>\%
 addScaleBar(options = list(imperial = FALSE)) \%>\%
 addPolygons(data=polygon_spdf,weight=1,col = 'white',
             opacity = 1, fillOpacity = 0) \%>\%
 addProviderTiles(providers$Esri.WorldImagery)
}
=======
% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clipLevel1BGeo.R
\name{clipxGeometry}
\alias{clipxGeometry}
\title{Clip GEDI x data by geometry}
\usage{
clipxGeometry(x, polygon_spdf, split_by = "id")
}
\arguments{
\item{x}{x; S4 object of class "gedi.level1b.dt"}

\item{polygon_spdf}{SpatialDataFrame. A polygon dataset for clipping the waveform}
}
\value{
Returns An object of class "gedi.level1b.dt"; subset of GEDI Level1B data
}
\description{
Clip GEDI x data within a given geometry area
}
\examples{

#' GEDI level1B file path
level1Bfilepath = system.file("extdata", "lvis_level1_clip.h5", package="rGEDI")

#' Reading GEDI level1B file
level1b = readLevel1b(level1Bfilepath)

#' Creating GEDI x object
x = x(level1b)

# Polgons file path
polygon_filepath <- system.file("extdata", "clip_polygon.shp", package="rGEDI")

# Reading GEDI level2B file
polygon_spdf<-raster::shapefile(polygons_filepath)

clipped_x = clipLevel1Geometry(x, polygon_spdf)

library(leaflet)
leaflet() \%>\%
 addCircleMarkers(clipped_x@dt$longitude_bin0,
                  clipped_x@dt$latitude_bin0,
                  radius = 1,
                  opacity = 1,
                  color = "red")  \%>\%
 addScaleBar(options = list(imperial = FALSE)) \%>\%
 addPolygons(data=polygon_spdf,weight=1,col = 'white',
             opacity = 1, fillOpacity = 0) \%>\%
 addProviderTiles(providers$Esri.WorldImagery)
}
>>>>>>> caio
