% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.gedi.R
\name{close,gedi.level2b-method}
\alias{close,gedi.level2b-method}
\title{Safely closes the \code{\linkS4class{gedi.level2b}}}
\usage{
\S4method{close}{gedi.level2b}(con, ...)
}
\arguments{
\item{con}{An object of class \code{gedi.level2b}}

\item{...}{Inherited from base}
}
\description{
Closing files will avoid locking HDF5 GEDI files.
}
