% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloud.R
\name{close.gedi.cloud_h5}
\alias{close.gedi.cloud_h5}
\title{Close a cloud HDF5 connection}
\usage{
\method{close}{gedi.cloud_h5}(con, ...)
}
\arguments{
\item{con}{A \code{gedi.cloud_h5} object returned by \code{\link[=readGEDICloud]{readGEDICloud()}}.}

\item{...}{Reserved for compatibility with \code{\link[=close]{close()}}.}
}
\value{
Invisibly returns \code{NULL}.
}
\description{
Close a cloud HDF5 connection
}
