% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdalBindings.R
\name{createDataset}
\alias{createDataset}
\title{Creates a new GDALDataset}
\usage{
createDataset(
  raster_path,
  nbands,
  datatype,
  projstring,
  lr_lat,
  ul_lat,
  ul_lon,
  lr_lon,
  res,
  nodata,
  co = NULL
)
}
\arguments{
\item{raster_path}{Character. The output path for the raster data}

\item{nbands}{Integer. Number of bands. Default 1.}

\item{datatype}{GDALDataType. The GDALDataType to use for the raster, use (GDALDataType$) to find the options. Default GDALDataType$GDT_Float64}

\item{projstring}{The projection string, either proj or WKT is accepted.}

\item{lr_lat}{Numeric. The lower right latitude.}

\item{ul_lat}{Numeric. The upper left latitude.}

\item{ul_lon}{Numeric. The upper left longitude.}

\item{lr_lon}{Numeric. The lower right longitude.}

\item{res}{Numeric. The resolution of the output raster}

\item{nodata}{Numeric. The no data value for the raster. Default 0.}

\item{co}{CharacterVector. A CharacterVector of creation options for GDAL. Default NULL}
}
\value{
An object from GDALDataset R6 class.
}
\description{
Create a new raster file based on specified data. It will output a *.tif file.
}
