% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clipLevel3.R
\name{crop,gedi.level3,numeric-method}
\alias{crop,gedi.level3,numeric-method}
\title{Crop gedi.level3 using by extents}
\usage{
\S4method{crop}{gedi.level3,numeric}(x, y, xb, xc, xd)
}
\arguments{
\item{x}{A \code{\linkS4class{gedi.level3}} object}

\item{y}{numeric. The lowest longitude xmin}

\item{xb}{numeric. The maximum longitude xmax}

\item{xc}{numeric. The mininum latitude ymin}

\item{xd}{numeric. The maximum latitude ymax}
}
\value{
The cropped gedi.level3
}
\description{
Crop gedi.level3 using by extents
}
