% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clipLevel3.R
\name{crop,gedi.level3,SpatVector-method}
\alias{crop,gedi.level3,SpatVector-method}
\title{Crop gedi.level3 using by extents}
\usage{
\S4method{crop}{gedi.level3,SpatVector}(x, y, xb = NULL, xc = NULL, xd = NULL)
}
\arguments{
\item{x}{A \code{\linkS4class{gedi.level3}} object}

\item{y}{SpatVect object from loaded with \code{\link[terra:vect]{terra::vect}}.}

\item{xb}{Not used.}

\item{xc}{Not used.}

\item{xd}{Not used.}
}
\value{
The cropped gedi.level3
}
\description{
Crop gedi.level3 using by extents
}
\examples{
level3path <- "C:/Users/caiohamamura/Downloads/GEDI03_rh100_mean_2019108_2020287_002_02.tif"
level3 <- readLevel3(level3path = level3path)


# Specifying the path to shapefile
polygon_filepath <- system.file("extdata", "stands_cerrado.shp", package = "rGEDI")

# Reading shapefile as SpatialPolygonsDataFrame object
library(terra)
polygon <- terra::vect(polygon_filepath)

# Clipping GEDI Full Waveform Geolocations by Geometry
level3_clip <- crop(level3, polygon, "id")

}
