% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gee.R
\name{ee_build_AlphaEarth_embedding_terrain_stack}
\alias{ee_build_AlphaEarth_embedding_terrain_stack}
\title{Build an AlphaEarth and terrain predictor stack}
\usage{
ee_build_AlphaEarth_embedding_terrain_stack(
  geom,
  start_year,
  end_year,
  add_lonlat = TRUE
)
}
\arguments{
\item{geom}{Area of interest.}

\item{start_year, end_year}{Inclusive year range.}

\item{add_lonlat}{Add coordinate bands.}
}
\value{
An Earth Engine image.
}
\description{
Build an AlphaEarth and terrain predictor stack
}
