% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gee.R
\name{ee_build_hls_s1c_terrain_stack}
\alias{ee_build_hls_s1c_terrain_stack}
\title{Build an HLS and terrain predictor stack in Earth Engine}
\usage{
ee_build_hls_s1c_terrain_stack(
  x,
  start_date,
  end_date,
  cloud_max = 20,
  buffer_m = 30
)
}
\arguments{
\item{x}{Area of interest.}

\item{start_date, end_date}{Date range.}

\item{cloud_max}{Maximum HLS cloud cover.}

\item{buffer_m}{AOI buffer in meters.}
}
\value{
An Earth Engine image.
}
\description{
Build an HLS and terrain predictor stack in Earth Engine
}
