% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gee.R
\name{ee_check_task_status}
\alias{ee_check_task_status}
\title{Check an Earth Engine task}
\usage{
ee_check_task_status(task, quiet = TRUE)
}
\arguments{
\item{task}{Earth Engine batch task.}

\item{quiet}{Logical.}
}
\value{
Task status list.
}
\description{
Check an Earth Engine task
}
