% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gee.R
\name{ee_initialize}
\alias{ee_initialize}
\title{Initialize Google Earth Engine}
\usage{
ee_initialize(
  project = Sys.getenv("EE_PROJECT", unset = ""),
  authenticate = TRUE,
  quiet = FALSE
)
}
\arguments{
\item{project}{Google Cloud project ID.}

\item{authenticate}{Authenticate when initialization fails.}

\item{quiet}{Logical.}
}
\value{
Invisibly returns the Python Earth Engine module.
}
\description{
Initialize Google Earth Engine
}
