% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gee.R
\name{ee_rect_to_sf}
\alias{ee_rect_to_sf}
\title{Convert an Earth Engine rectangle to \code{sf}}
\usage{
ee_rect_to_sf(aoi)
}
\arguments{
\item{aoi}{Earth Engine geometry.}
}
\value{
An \code{sf} polygon.
}
\description{
Convert an Earth Engine rectangle to \code{sf}
}
