% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gee.R
\name{ext_to_ee}
\alias{ext_to_ee}
\title{Convert an extent to an Earth Engine geometry}
\usage{
ext_to_ee(x)
}
\arguments{
\item{x}{A \code{SpatExtent}, raster/vector object, or numeric extent in
\code{c(xmin, xmax, ymin, ymax)} order.}
}
\value{
An Earth Engine geometry.
}
\description{
Convert an extent to an Earth Engine geometry
}
