% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gee.R
\name{extractEE}
\alias{extractEE}
\alias{extractGEDIAncillary}
\title{Extract Earth Engine image values at GEDI footprints}
\usage{
extractEE(stack, geom, scale = 30, chunk_size = 1000L)

extractGEDIAncillary(stack, geom, scale = 30, chunk_size = 1000L)
}
\arguments{
\item{stack}{Earth Engine image or list of images.}

\item{geom}{GEDI table, \code{sf}, or \code{SpatVector} points.}

\item{scale}{Sampling scale in meters.}

\item{chunk_size}{Number of points per request.}
}
\value{
A \link[data.table:data.table]{data.table::data.table}.
}
\description{
Extract Earth Engine image values at GEDI footprints
}
