% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rasterProducts.R
\name{extractLevel4B}
\alias{extractLevel4B}
\title{Extract GEDI Level 4B raster values}
\usage{
extractLevel4B(x, y, ...)
}
\arguments{
\item{x}{A raster path or \code{SpatRaster}.}

\item{y}{Points or polygons accepted by \code{\link[terra:vect]{terra::vect()}}.}

\item{...}{Passed to \code{\link[terra:extract]{terra::extract()}}.}
}
\value{
A data frame.
}
\description{
Extract GEDI Level 4B raster values
}
