% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdalBindings.R
\docType{data}
\name{GDALDataType}
\alias{GDALDataType}
\title{List of datatypes supported by the GDALDataset R6 class}
\format{
An object of class \code{list} of length 7.
}
\usage{
GDALDataType
}
\description{
List of datatypes supported by the GDALDataset R6 class
}
\keyword{datasets}
