% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.gedi.R
\docType{class}
\name{gedi.fullwaveform-class}
\alias{gedi.fullwaveform-class}
\alias{gedi.fullwaveform}
\title{Class for GEDI level1B Full Waveform}
\description{
Class for GEDI level1B Full Waveform
}
\section{Slots}{

\describe{
\item{\code{dt}}{Object of class data.table from \emph{data.table} package containing
the extracted GEDI full-waveform elevation and amplitude.}
}}

