% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.gedi.R
\docType{class}
\name{gedi.level1b.dt-class}
\alias{gedi.level1b.dt-class}
\title{Class for GEDI Level1B derived geolocation}
\description{
Class for GEDI Level1B derived geolocation
}
\section{Slots}{

\describe{
\item{\code{dt}}{Object of class data.table}
}}

