% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.gedi.R
\docType{class}
\name{gedi.level1bSim-class}
\alias{gedi.level1bSim-class}
\alias{gedi.level1bSim}
\title{Class for GEDI Full-Waveform Simulation}
\description{
Class for GEDI Full-Waveform Simulation
}
\section{Slots}{

\describe{
\item{\code{h5}}{Object of class H5File from hdf5r package}
}}

