% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.gedi.level1b.R
\docType{class}
\name{gedi.level1bSPDF-class}
\alias{gedi.level1bSPDF-class}
\title{Class for spatial points data frame from GEDI}
\description{
Class for spatial points data frame from GEDI
}
\section{Slots}{

\describe{
\item{\code{spdf}}{Object of class SpatialPointsDataFrame}
}}

