% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.gedi.R
\docType{class}
\name{gedi.level3-class}
\alias{gedi.level3-class}
\alias{gedi.level3}
\title{Class for GEDI level3}
\description{
Class for GEDI level3
}
\section{Slots}{

\describe{
\item{\code{raster}}{Object of class \code{\link[terra:SpatRaster-class]{SpatRaster}} from \code{terra} package containing the
GEDI level3 products: rasterized metrics}
}}

\seealso{
\url{https://daac.ornl.gov/GEDI/guides/GEDI_L3_LandSurface_Metrics_V2.html}
}
