% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C_gediWFCollocate.R
\name{gediWFCollocate}
\alias{gediWFCollocate}
\title{Collocate waveforms based in ALS}
\usage{
gediWFCollocate(
  output,
  als,
  gedi,
  aEPSG,
  readHDFgedi = TRUE,
  lgw = FALSE,
  readHDFlvis = FALSE,
  bounds = NULL,
  leaveEmpty = NULL,
  lEPSG = 4326,
  beamList = NULL,
  skipBeams = NULL,
  readBeams = NULL,
  solveCofG = NULL,
  maxVshift = -10,
  vStep = 0.2,
  maxShift = 10,
  step = 1,
  simplex = FALSE,
  anneal = FALSE,
  fixFsig = FALSE,
  geoError = FALSE,
  quickGeo = FALSE,
  optTol = 0.2,
  maxIter = 300,
  writeSimProg = NULL,
  writeWaves = NULL,
  nTriesAnneal = NULL,
  itersFixedT = NULL,
  kAnneal = NULL,
  tInitial = NULL,
  muAnneal = NULL,
  tMinAnneal = NULL,
  hOffset = NULL,
  offset = NULL,
  fSigma = NULL,
  pSigma = NULL,
  readPulse = NULL,
  minSense = NULL,
  maxZen = NULL,
  maxScanAng = NULL,
  minDense = NULL,
  decimate = NULL,
  noFilt = NULL,
  filtOut = NULL,
  smooth = NULL,
  checkCover = NULL,
  median = NULL,
  noNorm = NULL,
  norm = NULL,
  allSimMeth = NULL,
  pulseBefore = NULL,
  noOctree = NULL,
  octLevels = NULL,
  nOctPix = NULL
)
}
\arguments{
\item{output}{output filename}

\item{als}{file;        input als file}

\item{gedi}{file;       single input GEDI/LVIS L1B file}

\item{aEPSG}{epsg;      ALS projection}

\item{bounds}{minX minY maxX maxY;    bounds to use, specified in ALS projection}

\item{lEPSG}{epsg;      LVIS projection}

\item{beamList}{11111111; 0/1 for whether or not to use beams 1-8}

\item{skipBeams}{n;     list of beam numbers to skip. No spaces between (eg 123)}

\item{readBeams}{n;     list of beam numbers to read. No spaces between (eg 123)}

\item{solveCofG}{use centre of gravity to match vertical offsets}

\item{maxVshift}{x;     grid or geoError mode, vertical distance to search over}

\item{vStep}{z;         grid or geoError mode, vertical step size}

\item{maxShift}{x;      grid mode, horizontal distance to search over}

\item{step}{x;          grid mode, horizontal step size}

\item{simplex}{use simplex optimisation rather than doing the full bullseye plot}

\item{anneal}{use simulated annealing optimisation}

\item{fixFsig}{fix fSigma in simplex}

\item{geoError}{expError correlDist; rapid geolocation, using expected geolocation
error and correlation distance. Vertical shifts must be separatley defined}

\item{quickGeo}{perform rapid geolocation using default error values}

\item{optTol}{x;        tolerance for optimisation}

\item{maxIter}{n;       maximum number of iterations}

\item{writeSimProg}{write progress of simplex to output}

\item{writeWaves}{name; write out final waveforms as HDF5 when using simplex}

\item{nTriesAnneal}{n;  how many points do we try before stepping?}

\item{itersFixedT}{n;   how many iterations for each T?}

\item{kAnneal}{x;       Boltzmann constant for annealing}

\item{tInitial}{x; A initial annealing temperature}

\item{muAnneal}{x;      damping factor for temperature}

\item{tMinAnneal}{x;    minimum annealing temperature}

\item{hOffset}{dx dy;   centre of horizontal offsets}

\item{offset}{z;        vertical datum offset

Waveform characteristics}

\item{fSigma}{x;        footprint width, sigma in metres}

\item{pSigma}{x;        Gaussian pulse length, sigma in metres}

\item{readPulse}{file;  pulse shape, if not Gaussian}

\item{minSense}{x;      minimum LVIS/GEDI beam sensitivity to accept}

\item{maxZen}{zen;      maximum LVIS/GEDI zenith angle to use, degrees}

\item{maxScanAng}{ang;  maximum ALS scan angle, degrees}

\item{minDense}{x;      minimum ALS beam density to accept}

\item{decimate}{f;      decimate ALS point cloud by a factor, to save RAM}

\item{noFilt}{don't filter outliers from correlation (default)}

\item{filtOut}{s;       filter outliers from correlation stats, along with
the number of standard deviations to use as a threshold}

\item{smooth}{sig;      smooth both waves before comparing}

\item{checkCover}{only include footprints that are at least 2/3 covered with ALS data}

\item{median}{use median correlation rather than mean}

\item{noNorm}{don't correct sims for ALS densiy variations}

\item{norm}{correct sims for ALS densiy variations}

\item{allSimMeth}{use all simulation methods}

\item{pulseBefore}{apply pulse shape before binning to prevent aliasing}

\item{noOctree}{do not use an octree}

\item{octLevels}{n;    number of octree levels to use}

\item{nOctPix}{n;      number of octree pixels along a side for the top level}

\item{readHDFgedi;}{read GEDI HDF5 input (default)}

\item{lgw;}{LVIS is in lgw (default is GEDI hdf5)}

\item{readHDFlvis;}{read GEDI HDF5 input (default is GEDI hdf5)}

\item{leaveEmpty;}{exit if there are no usable footprints}
}
\value{
Returns an S4 object of class \code{\link[hdf5r:H5File-class]{hdf5r::H5File}}
containing the collocated waves
}
\description{
Uses the correlation method in Blair and Hofton (1999) to colocate a large-footprint
lidar dataset with a small-footprint, discrete-return dataset. Note that it requires the full-waveform
LVIS or GEDI data, which is contained in the L1B files. It uses the Pearson correlation to find the best
affine transformation (x and y only, or x, y and z) and footprint size needed to align a large-footprint
dataset with a small-footprint dataset. It has three potential modes of operation.
\itemize{
\item It can test a grid of affine transformations and give the correlation for every point
(for a single footprint width), as used in Blair and Hofton (1999).
\item It can use a simplex to move along the error surface and find the optimum transformation
and footprint size. Note that initial location needs to be within around 20 m of the true location
for their to be a sufficient gradient on the error surface.
\item It can use a hybrid of the two, testing every step on a coarse grid, then setting a simplex off
from the location of maximum correlation. This allows a rapid assessment to get within 20 m of the
true value, then uses the simplex to find the offsets precisely.
\item It requires the large-footprint system pulse shape (either a Gaussian width or a file containing
range and intensity) and the EPSG codes for the two datasets. It reads ALS data in .las format and
can read either simulated HDF5 files from gediRat or LVIS in HDF5 or lgw format. A reader for GEDI
data will be added once that data is available.
}

If the full grid is used, it outputs an ASCII file with the correlation for each x, y and z offset.
If it uses a simplex it outputs the single optimum offset.
}
\examples{
\dontshow{
rm(list = ls())
}

}
\seealso{
i) Hancock, S., Armston, J., Hofton, M., Sun, X., Tang, H., Duncanson, L.I., Kellner,
J.R. and Dubayah, R., 2019. The GEDI simulator: A large-footprint waveform lidar simulator
for calibration and validation of spaceborne missions. Earth and Space Science.
\doi{10.1029/2018EA000506}

ii) gediSimulator: \url{https://bitbucket.org/StevenHancock/gedisimulator/src/master/}
}
