% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C_addNoise.R
\name{gediWFNoise}
\alias{gediWFNoise}
\title{GEDI add noise to simulated waveform}
\usage{
gediWFNoise(input, output, seed = NULL, linkNoise = NULL,
  dcBias = NULL, linkFsig = NULL, linkPsig = NULL, trueSig = NULL,
  bitRate = NULL)
}
\arguments{
\item{input}{\code{\link{character}}. Waveform  input filename}

\item{output}{\code{\link{character}}. Output filename}

\item{seed}{\code{\link{integer}}. Random number seed}

\item{linkNoise}{\code{\link{numeric}(2)}. Apply Gaussian noise based on link margin at a cover}

\item{dcBias}{\code{\link{numeric}}. Mean noise level}

\item{linkFsig}{\code{\link{numeric}}. Footprint width to use when calculating and applying signal noise}

\item{linkPsig}{\code{\link{numeric}}. Pulse width to use when calculating and applying signal noise}

\item{trueSig}{\code{\link{numeric}}. True sigma of background noise}

\item{bitRate}{\code{\link{integer}}. DN bit rate}
}
\description{
Add noise to HDF output from gediWFSimulator
}
