% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clipLevel3.R
\name{geom_spatraster,ANY,gedi.level3-method}
\alias{geom_spatraster,ANY,gedi.level3-method}
\title{Maps geom_spatraster function from tidyterra}
\usage{
\S4method{geom_spatraster}{ANY,gedi.level3}(mapping = tidyterra::aes(), data, ...)

\S4method{geom_spatraster}{ANY,gedi.level3}(mapping = tidyterra::aes(), data, ...)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by \code{\link[ggplot2:aes]{ggplot2::aes()}} or
\code{\link[ggplot2:aes]{ggplot2::aes()}}. See Aesthetics specially in the use of fill aesthetic.}

\item{data}{A \code{\linkS4class{gedi.level3}} object}

\item{...}{Other arguments passed on to \code{\link[ggplot2:layer]{ggplot2::layer()}}. These are often
aesthetics, used to set an aesthetic to a fixed value, like colour = "red" or
size = 3. They may also be parameters to the paired geom/stat.}
}
\value{
A ggplot2 layer

A ggplot2 layer
}
\description{
Creates the ggplot2 layer for gedi.level3
}
\seealso{
\code{\link[tidyterra:geom_spatraster]{tidyterra::geom_spatraster}}
}
