% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clipLevel3.R
\name{geom_spatraster}
\alias{geom_spatraster}
\title{Generic function for geom_spatraster}
\usage{
geom_spatraster(mapping = tidyterra::aes(), data, ...)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by \code{\link[ggplot2:aes]{ggplot2::aes()}} or
\code{\link[ggplot2:aes]{ggplot2::aes()}}. See Aesthetics specially in the use of fill aesthetic.}

\item{data}{A \code{SpatRaster} object}

\item{...}{Other arguments passed on to \code{\link[ggplot2:layer]{ggplot2::layer()}}. These are often
aesthetics, used to set an aesthetic to a fixed value, like colour = "red" or
size = 3. They may also be parameters to the paired geom/stat.}
}
\value{
A ggplot2 layer
}
\description{
Generic function for geom_spatraster
}
\examples{
level3path <- "C:/Users/caiohamamura/Downloads/GEDI03_rh100_mean_2019108_2020287_002_02.tif"
level3 <- readLevel3(level3path = level3path)
xmin <- -5888627
xmax <- -5878615
ymin <- -990000
ymax <- -976400

# Clipping GEDI Full Waveform Geolocations by boundary box extent
level3_crop <- crop(level3, xmin, xmax, ymin, ymax)

if (require(ggplot2)) {
  ggplot() +
    geom_spatraster(data = level3_crop)
}
}
\seealso{
\code{\link[tidyterra:geom_spatraster]{tidyterra::geom_spatraster}}
}
