% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling.R
\name{geomSampling}
\alias{geomSampling}
\title{GEDI polygon-stratified sampling specification}
\usage{
geomSampling(size, geom, split_id = NULL, lon = NULL, lat = NULL)
}
\arguments{
\item{size}{Number or fraction per polygon.}

\item{geom}{Polygon object.}

\item{split_id}{Optional polygon ID column.}

\item{lon, lat}{Optional coordinate column names.}
}
\description{
GEDI polygon-stratified sampling specification
}
