% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gee.R
\name{get_catalog_id}
\alias{get_catalog_id}
\title{Return an Earth Engine catalog ID}
\usage{
get_catalog_id(x)
}
\arguments{
\item{x}{A GEDI product name, catalog search row, or catalog ID.}
}
\value{
A character catalog ID.
}
\description{
Return an Earth Engine catalog ID
}
