% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gee.R
\name{getTileUrl}
\alias{getTileUrl}
\title{Get an Earth Engine image tile URL}
\usage{
getTileUrl(image, vis = list())
}
\arguments{
\item{image}{Earth Engine image.}

\item{vis}{Visualization parameter list.}
}
\value{
Character tile URL.
}
\description{
Get an Earth Engine image tile URL
}
