% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLevel1BWF.R
\name{getxWF}
\alias{getxWF}
\title{Get GEDI Level1B waveform}
\usage{
getxWF(x, shot_number)
}
\arguments{
\item{shot_number}{Shot number, an vector represeting}

\item{level1b}{h5file; S4 object of class H5File}
}
\value{
Returns An object of class "gedi.waveform";
}
\description{
Extract GEDI Level1B waveform for a given shot number
}
\examples{

#' LVIS level 2 file path
level1_filepath = system.file("extdata", "lvis_level1_clip.h5", package="rLVIS")

# Rectangle
xleft = 9.35986
xright = 9.35988
ybottom = 0.5786
ytop = 0.5790

#' Reading LVIS level 2 file
level1_waveform = readLevel1b(level1_filepath)

output = tempfile(fileext="h5")

clipped_waveform = clipLevel1(level1_waveform, output, xleft, xright, ybottom, ytop)

}
