% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gee.R
\name{glcmTexture}
\alias{glcmTexture}
\title{Earth Engine gray-level co-occurrence textures}
\usage{
glcmTexture(x, size = 1L, kernel = NULL, average = TRUE)
}
\arguments{
\item{x}{Earth Engine integer image.}

\item{size}{Neighborhood size.}

\item{kernel}{Optional Earth Engine kernel.}

\item{average}{Average directional metrics.}
}
\value{
An Earth Engine image.
}
\description{
Earth Engine gray-level co-occurrence textures
}
