% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling.R
\name{gridSampling}
\alias{gridSampling}
\title{GEDI grid-stratified sampling specification}
\usage{
gridSampling(size, grid_size, lon = NULL, lat = NULL)
}
\arguments{
\item{size}{Number or fraction per cell.}

\item{grid_size}{Grid size in decimal degrees.}

\item{lon, lat}{Optional coordinate column names.}
}
\description{
GEDI grid-stratified sampling specification
}
