% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/level4A.R
\name{gridStatsLevel4A}
\alias{gridStatsLevel4A}
\title{Grid GEDI Level 4A footprint metrics}
\usage{
gridStatsLevel4A(level4A, metric = "agbd", fun = mean, res = 0.01, ...)
}
\arguments{
\item{level4A}{A table returned by \code{\link[=getLevel4A]{getLevel4A()}}.}

\item{metric}{Metric column name.}

\item{fun}{Aggregation function.}

\item{res}{Output resolution in decimal degrees.}

\item{...}{Additional arguments passed to \code{fun}.}
}
\value{
A \code{\link[terra:SpatRaster-class]{terra::SpatRaster}}.
}
\description{
Grid GEDI Level 4A footprint metrics
}
