% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{level1b.dt}
\alias{level1b.dt}
\title{Read LVIS Level1b data}
\usage{
level1b.dt(level1b)
}
\arguments{
\item{level1b}{H5File object from level1b}
}
\value{
H5File; S4 object of class H5File;
}
\description{
This function reads LVIS level1b data
}
\seealso{
\code{\link[h5]{h5file}} in the \emph{h5} package.
}
\author{
Carlos Alberto Silva. This function calls \emph{h5file} function from h5 package (Author: Mario Annau)
}
