% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/level1B2dt.R
\name{level1B2dt}
\alias{level1B2dt}
\title{Read LVIS Level1b data}
\usage{
level1B2dt(level1b, select = c("latitude_bin0", "latitude_lastbin",
  "longitude_bin0", "longitude_lastbin", "shot_number"))
}
\arguments{
\item{level1b}{H5File object from level1b}

\item{select}{character vector of desired columns to extract from GEDI hdf5 format, default c("latitude_bin0", "latitude_lastbin", "longitude_bin0", "longitude_lastbin", "shot_number")}
}
\value{
gedi.level1b class; S4 object of class gedi.level1b;
}
\description{
This function reads LVIS level1b data
}
\seealso{
\code{\link[h5]{h5file}} in the \emph{h5} package.
}
\author{
Carlos Alberto Silva. This function calls \emph{h5file} function from h5 package (Author: Mario Annau)
}
