% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/level2AGridStats.R
\name{level2AGridStats}
\alias{level2AGridStats}
\title{#' Computing a serie of LVIS metrics
mlvis<-GridMetrics(level2a.dt=level2a.dt,func=mySetOfMetrics(ZT), res=0.0005)}
\usage{
level2AGridStats(x, func = max(rh100), res = 0.5)
}
\description{
#' Computing the maximum of RH100 only
maxRH100<-GridMetrics(level2a.dt=level2a.dt,func=max(RH100), res=0.0005)
}
