% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/level2BRasterize.R
\name{level2bRasterizeStats}
\alias{level2bRasterizeStats}
\title{Aggregate selected metrics into raster tif cells}
\usage{
level2bRasterizeStats(
  l2bDir,
  metrics,
  out_root,
  ul_lat,
  ul_lon,
  lr_lat,
  lr_lon,
  res,
  creation_options = def_co,
  agg_function = default_agg_function,
  agg_join = default_agg_join,
  finalizer = default_finalizer
)
}
\arguments{
\item{l2bDir}{CharacterVector. The directory paths where the H5 GEDI files are stored;}

\item{metrics}{CharacterVector. A vector of metrics available from Level2B, as in the getLevel2BVPM documentation}

\item{out_root}{Character. The root name for the raster output files, the pattern is {out_root}_{metric}_{count/m1/m2/m3/m4}.tif. This should include the full path for the file.}

\item{ul_lat}{Numeric. Upper left latitude for the bounding box}

\item{ul_lon}{Numeric. Upper left longitude for the bounding box}

\item{lr_lat}{Numeric. Lower right latitude for the bounding box}

\item{lr_lon}{Numeric. Lower right longitude for the bounding box}

\item{res}{NumericVector. Resolution lon lat for the output raster in coordinates decimal degrees}

\item{creation_options}{CharacterVector. The GDAL creation options for the tif file. Default c("COMPRESS=PACKBITS", "BIGTIFF=IF_SAFER", "TILED=YES", "BLOCKXSIZE=512", "BLOCKYSIZE=512") will create BIGTIFF if needed, with DEFLATE compression and tiled by 512x512 pixels.}

\item{agg_function}{Function. An aggregate function which should return a data.table with the aggregate statistics}

\item{agg_join}{Function. A function to merge two different agg objects.}

\item{finalizer}{List<name, formula>. A list with the final raster names and the formula which uses the base statistics.}
}
\value{
Nothing. It outputs multiple raster tif files to the out_root specified path.
}
\description{
This function will read multiple Level2B H5 files and aggregate into multiple rasters: count, and 1st, 2nd, 3rd and 4th moments (count, m1, m2, m3 and m4) for each metric selected, from which we can calculate statistics such as Mean, SD, Skewness and Kurtosis.
}
\details{

}
\examples{
# Specifying the path to GEDI level2B data (zip file)
library(rGEDI)
library(data.table)

outdir = tempdir()
level2B_fp_zip <- system.file("extdata",
                   "GEDI02_B_2019108080338_O01964_T05337_02_001_01_sub.zip",
                   package="rGEDI")

# Unzipping GEDI level2A data
level2Bpath <- unzip(level2B_fp_zip,exdir = outdir)

# Reading GEDI level2B data (h5 file)
level2b<-readLevel2B(level2Bpath = level2Bpath)

ul_lat = -13.72016
ul_lon = -44.14000
lr_lat = -13.74998
lr_lon = -44.11009

res = 100 # meters
lat_to_met_factor = 1 / 110540
lon_to_met_factor = 1 / 111320
xres = lon_to_met_factor * res
yres = lat_to_met_factor * res

agg_function = function(x) data.table(
    min = min(x), 
    max = max(x), 
    sum = sum(x), 
    n = length(x))

agg_join = function(agg1, agg2) {
agg1[is.na(agg1)] = 0
data.table(
    min = pmin(agg1$min, agg2$min), 
    max = pmax(agg1$max, agg2$max), 
    sum = agg1$sum + agg2$sum, 
    n = agg1$n + agg2$n
) 
}

finalizer = list(
    mean = "sum/n",
    range = "max-min"
)

level2bRasterizeStats(
  outdir,
  metrics = c("rh100"),
  out_root = file.path(outdir, "output"),
  ul_lat = ul_lat,
  ul_lon = ul_lon,
  lr_lat = lr_lat,
  lr_lon = lr_lon,
  res = c(xres, -yres),
  creation_options = c("COMPRESS=DEFLATE" ,
    "BIGTIFF=IF_SAFER",
    "TILED=YES",
    "BLOCKXSIZE=512",
    "BLOCKYSIZE=512"),
  agg_function = agg_function,
  agg_join = agg_join,
  finalizer = finalizer
  )

close(level2b)

}
\references{
\insertAllCited{}

Terriberry, Timothy B. (2007), Computing Higher-Order Moments Online, archived from the original on 23 April 2014, retrieved 5 May 2008
}
