% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/level2BVPMGridStats.R
\name{level2BVPMGridStats}
\alias{level2BVPMGridStats}
\title{#' Computing a serie of LVIS metrics
mlvis<-GridMetrics(level2a.dt=level2a.dt,func=mySetOfMetrics(ZT), res=0.0005)}
\usage{
level2BVPMGridStats(x, func, res = 0.5)
}
\description{
#' Computing the maximum of RH100 only
maxRH100<-GridMetrics(level2a.dt=level2a.dt,func=max(RH100), res=0.0005)
}
