% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LPDAACDataPool.R
\name{LPDAACDataPool}
\alias{LPDAACDataPool}
\title{Download GEDI data}
\usage{
LPDAACDataPool(filepath,outdir)
}
\arguments{
\item{filepath}{Vector object; path to the GEDI data}

\item{outdir}{Vector object, output directory for downloading GEDI data}
}
\description{
Download GEDI data from LP DAAC Data Pool. Users will need to enter their
Earth Explore login Information for downloading the data.
}
\examples{
\dontrun{
#' Set path to GEDI data
filepath=c(paste0(
                 "https://e4ftl01.cr.usgs.gov/GEDI/GEDI02_B.001",
                 "/2019.04.18/GEDI02_B_2019108032534_O01961_T03911_02_001_01.h5"
                 ),
          paste0("https://e4ftl01.cr.usgs.gov/GEDI/GEDI02_B.001",
                 "/2019.04.18/GEDI02_B_2019108045815_O01962_T01066_02_001_01.h5"
                )
         )

# Set dir to download files to
outdir=tempdir()

#' Downloading GEDI data
LPDAACDataPool(filepath,outdir)
}
}
\references{
Credits to Cole Krehbiel. Code adpted from <https://git.earthdata.nasa.gov/projects/LPDUR/repos/daac_data_download_r/browse/DAACDataDownload.R>
}
