% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gee.R
\name{map_create}
\alias{map_create}
\title{Apply an Earth Engine model to a predictor image}
\usage{
map_create(model, stack, aoi = NULL, name = "prediction")
}
\arguments{
\item{model}{A trained Earth Engine classifier or regressor.}

\item{stack}{Earth Engine predictor image.}

\item{aoi}{Optional area to clip.}

\item{name}{Output band name.}
}
\value{
An Earth Engine image.
}
\description{
Apply an Earth Engine model to a predictor image
}
