% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gee.R
\name{map_download}
\alias{map_download}
\title{Download an Earth Engine image}
\usage{
map_download(
  image,
  filename,
  region,
  scale = 30,
  crs = "EPSG:4326",
  overwrite = FALSE
)
}
\arguments{
\item{image}{Earth Engine image.}

\item{filename}{Destination GeoTIFF path.}

\item{region}{Area of interest.}

\item{scale}{Pixel size in meters.}

\item{crs}{Output CRS.}

\item{overwrite}{Logical.}
}
\value{
Invisibly returns \code{filename}.
}
\description{
Download an Earth Engine image
}
