% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gee.R
\name{map_view}
\alias{map_view}
\title{Display Earth Engine layers in Leaflet}
\usage{
map_view(layers, vis = list(), aoi = NULL)
}
\arguments{
\item{layers}{Named list of Earth Engine images.}

\item{vis}{Named list of visualization parameter lists.}

\item{aoi}{Optional spatial object used to fit map bounds.}
}
\value{
A Leaflet widget.
}
\description{
Display Earth Engine layers in Leaflet
}
