% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rasterProducts.R
\name{mosaicLevel3}
\alias{mosaicLevel3}
\title{Mosaic GEDI Level 3 rasters}
\usage{
mosaicLevel3(x, fun = "mean", filename = "", overwrite = FALSE)
}
\arguments{
\item{x}{Raster paths, a \code{SpatRaster}, or a list of rasters.}

\item{fun}{Function used for overlapping cells.}

\item{filename}{Optional output filename.}

\item{overwrite}{Logical.}
}
\value{
A \code{\link[terra:SpatRaster-class]{terra::SpatRaster}}.
}
\description{
Mosaic GEDI Level 3 rasters
}
