% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rasterProducts.R
\name{mosaicLevel4B}
\alias{mosaicLevel4B}
\title{Mosaic GEDI Level 4B rasters}
\usage{
mosaicLevel4B(x, fun = "mean", filename = "", overwrite = FALSE)
}
\arguments{
\item{x}{Raster paths, a \code{SpatRaster}, or a list of rasters.}

\item{fun}{Function used for overlapping cells.}

\item{filename}{Optional output filename.}

\item{overwrite}{Logical.}
}
\value{
A \code{\link[terra:SpatRaster-class]{terra::SpatRaster}}.
}
\description{
Mosaic GEDI Level 4B rasters
}
