% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/level2AGridStats.R, R/level2BVPMGridStats.R,
%   R/level2BVPMStats.R
\name{mySetOfMetrics}
\alias{mySetOfMetrics}
\title{Compute a Series of Grid Metrics}
\usage{
mySetOfMetrics(x)

mySetOfMetrics(x)

L2Stats(level2_spdf, func, id)
}
\arguments{
\item{level2a.dt}{LVIS l2 dataset; object of class \code{SpatialPointsDataFrame}}

\item{func}{the function to be applied to each cell}

\item{res}{spatial resolution for the output raster layer}

\item{level2a.dt}{LVIS l2 dataset; object of class \code{SpatialPointsDataFrame}}

\item{func}{the function to be applied to each cell}

\item{res}{spatial resolution for the output raster layer}

\item{level2_spdf}{LVIS l2 dataset; object of class \code{SpatialPointsDataFrame}}

\item{func}{the function to be applied to each cell}

\item{id}{a vector contatining the id for each LVIS Level2 observation. Defaut is NULL}
}
\value{
Returns raster layer (s) of selected LVIS metric (s)

Returns raster layer (s) of selected LVIS metric (s)

Returns a \code{data.table} object containting Statistics of LVIS Level2-derived Metrics
}
\description{
This function computes a series of user-defined descriptive statistics for a LVIS dataset within
each grid cell

This function computes a series of user-defined descriptive statistics for a LVIS dataset within
each grid cell

Computes a Series of Statistics from LVIS Level2-derived Metrics
}
\author{
Carlos Alberto Silva (This function has been adapted from the grid_metrics function in lidR package. All credits to Roussel et al. 2019).

Carlos Alberto Silva (This function has been adapted from the grid_metrics function in lidR package. All credits to Roussel et al. 2019).

Carlos Alberto Silva (This function has been adapted from the grid_metrics function in lidR package. All credits to Roussel et al. 2019).
}
