% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clipLevel3.R
\name{plot,gedi.level3,missing-method}
\alias{plot,gedi.level3,missing-method}
\title{Plot function for gedi.level3 data}
\usage{
\S4method{plot}{gedi.level3,missing}(x, y, ...)
}
\arguments{
\item{x}{A \code{\linkS4class{gedi.level3}} object}

\item{y}{Not used, inherited by generic method}

\item{...}{Additional parameters for plotting}
}
\value{
Nothing
}
\description{
Plots gedi.level3 raster data
}
\examples{
level3path <- "C:/Users/caiohamamura/Downloads/GEDI03_rh100_mean_2019108_2020287_002_02.tif"
level3 <- readLevel3(level3path = level3path)
xmin <- -5888627
xmax <- -5878615
ymin <- -990000
ymax <- -976400

# Clipping GEDI Full Waveform Geolocations by boundary box extent
level3_crop <- crop(level3, xmin, xmax, ymin, ymax)

plot(level3_crop)

}
