% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeling.R
\name{plot.gedi_var_selection}
\alias{plot.gedi_var_selection}
\title{Plot GEDI variable selection results}
\usage{
\method{plot}{gedi_var_selection}(x, which = c("importance", "rfe"), ...)
}
\arguments{
\item{x}{A \code{gedi_var_selection} object.}

\item{which}{Plot variable importance or RFE error.}

\item{...}{Additional graphical parameters.}
}
\value{
The object, invisibly.
}
\description{
Plot GEDI variable selection results
}
