% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.gedi.R
\docType{methods}
\name{plot,gedi.level1bSim,missing-method}
\alias{plot,gedi.level1bSim,missing-method}
\title{Plot simulated GEDI full-waveform}
\usage{
\S4method{plot}{gedi.level1bSim,missing}(x, relative = FALSE,
  polygon = FALSE, method = "RXWAVEINT", ...)
}
\arguments{
\item{x}{An object of class "gedi.level1bSim". (output of \code{\link[rGEDI:gediWFSimulator]{gediWFSimulator}} function)}

\item{relative}{if TRUE, the Wavform Amplitude will be showed in percentage (\%)}

\item{polygon}{if TRUE, polygon will be added to the plot}

\item{method}{methods used for simulating the GEDI full-waveform ("RXWAVEINT","RXWAVEINT" or "RXWAVEINT"). Default is "RXWAVECOUNT".}

\item{...}{will be passed to the main plot}
}
\description{
Plots a single GEDI full-waveform (level1b) simulated from ALS 3D-point cloud
}
\examples{
lasfile_amazon <- system.file("extdata", "Amazon.las", package="rGEDI")
lasfile_cerrado <- system.file("extdata", "Cerrado.las", package="rGEDI")

# Reading and plot ALS file
libsAvailable = require(lidR) && require(plot3D)
if (libsAvailable) {
las_amazon<-readLAS(lasfile_amazon)
las_cerrado<-readLAS(lasfile_cerrado)

# Extracting plot center geolocations
xcenter_amazon = mean(las_amazon@bbox[1,])
ycenter_amazon = mean(las_amazon@bbox[2,])
xcenter_cerrado = mean(las_cerrado@bbox[1,])
ycenter_cerrado = mean(las_cerrado@bbox[2,])

# Simulating GEDI full-waveform
wf_amazon<-gediWFSimulator(
                          input=lasfile_amazon,
                          output=paste0(
                                        getwd(),
                                        "//gediWF_amazon_simulation.h5"
                                        ),
                          coords = c(xcenter_amazon, ycenter_amazon))
wf_cerrado<-gediWFSimulator(
                            input=lasfile_cerrado,
                            output=paste0(
                                          getwd(),
                                          "//gediWF_cerrado_simulation.h5"
                                          ),
                            coords = c(xcenter_cerrado, ycenter_cerrado))
# Plot Full-waveform
par(mfrow=c(2,2), mar=c(4,4,0,0), oma=c(0,0,1,1),cex.axis = 1.2)
scatter3D(
         las_amazon@data$X,
         las_amazon@data$Y,
         las_amazon@data$Z,
         pch = 16, colkey = FALSE, main="",
         cex = 0.5, bty = "u", col.panel ="gray90",
         phi = 30, alpha=1, theta=45, col.grid = "gray50", 
         xlab="UTM Easting (m)", ylab="UTM Northing (m)", zlab="Elevation (m)"
         )

plot(wf_amazon, relative=TRUE, polygon=TRUE, type="l", lwd=2, col="forestgreen",
    xlab="", ylab="Elevation (m)", ylim=c(90,140))
grid()
scatter3D(
         las_cerrado@data$X,las_cerrado@data$Y,las_cerrado@data$Z,
         pch = 16,colkey = FALSE, main="",
         cex = 0.5,bty = "u",col.panel ="gray90",
         phi = 30,alpha=1,theta=45,col.grid = "gray50", 
         xlab="UTM Easting (m)", ylab="UTM Northing (m)", zlab="Elevation (m)"
         )

plot(wf_cerrado, relative=TRUE, polygon=TRUE, type="l", lwd=2, col="green",
    xlab="Waveform Amplitude (\%)", ylab="Elevation (m)", ylim=c(815,835))
grid()
}
}
