% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rasterProducts.R
\name{plotLevel3}
\alias{plotLevel3}
\title{Plot GEDI Level 3 metrics}
\usage{
plotLevel3(x, ...)
}
\arguments{
\item{x}{A raster path or \code{SpatRaster}.}

\item{...}{Passed to \code{\link[terra:plot]{terra::plot()}}.}
}
\value{
Invisibly returns the raster.
}
\description{
Plot GEDI Level 3 metrics
}
