% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/level4A.R
\name{plotLevel4A}
\alias{plotLevel4A}
\title{Plot GEDI Level 4A footprint biomass}
\usage{
plotLevel4A(level4A, metric = "agbd", ...)
}
\arguments{
\item{level4A}{A table returned by \code{\link[=getLevel4A]{getLevel4A()}}.}

\item{metric}{Metric column to display.}

\item{...}{Additional arguments passed to \code{\link[graphics:plot]{graphics::plot()}}.}
}
\value{
Invisibly returns \code{level4A}.
}
\description{
Plot GEDI Level 4A footprint biomass
}
