% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rasterProducts.R
\name{plotLevel4B}
\alias{plotLevel4B}
\title{Plot GEDI Level 4B metrics}
\usage{
plotLevel4B(x, ...)
}
\arguments{
\item{x}{A raster path or \code{SpatRaster}.}

\item{...}{Passed to \code{\link[terra:plot]{terra::plot()}}.}
}
\value{
Invisibly returns the raster.
}
\description{
Plot GEDI Level 4B metrics
}
