% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdalBindings.R
\name{+.GDALRasterBand}
\alias{+.GDALRasterBand}
\title{GDALRasterBand + operation}
\usage{
\method{+}{GDALRasterBand}(first, second)

\method{+}{GDALRasterBand}(first, second)
}
\arguments{
\item{first}{GDALRasterBand. The raster band to process.}

\item{second}{Atomic or GDALRasterBand. The other object to operate.}

\item{x}{GDALDatset. Automatically obtained from GDALDataset[[]] call.}

\item{slice}{Integer. The index for the band to access.}
}
\value{
An object of GDALRasterBand R6 class.

An object of GDALRasterBand R6 class.
}
\description{
This function gives access to the GDALRasterBand using [[i]], where i is the band index to return.

This function gives access to the GDALRasterBand using [[i]], where i is the band index to return.
}
