% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/level4A.R
\name{polyStatsLevel4A}
\alias{polyStatsLevel4A}
\title{Polygon statistics for GEDI Level 4A footprints}
\usage{
polyStatsLevel4A(level4A, polygon, metric = "agbd", fun = mean, id = NULL, ...)
}
\arguments{
\item{level4A}{A table returned by \code{\link[=getLevel4A]{getLevel4A()}}.}

\item{polygon}{An \code{sf}, \code{sfc}, or \code{SpatVector} polygon object.}

\item{metric}{Metric column name.}

\item{fun}{Aggregation function.}

\item{id}{Optional polygon ID field.}

\item{...}{Additional arguments passed to \code{fun}.}
}
\value{
A data table with one row per polygon.
}
\description{
Polygon statistics for GEDI Level 4A footprints
}
