% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rasterProducts.R
\name{polyStatsLevel4B}
\alias{polyStatsLevel4B}
\title{Polygon summaries of GEDI Level 4B metrics}
\usage{
polyStatsLevel4B(x, y, fun = mean, na.rm = TRUE, ...)
}
\arguments{
\item{x}{A raster path or \code{SpatRaster}.}

\item{y}{Points or polygons accepted by \code{\link[terra:vect]{terra::vect()}}.}

\item{fun}{Summary function.}

\item{na.rm}{Logical.}

\item{...}{Passed to \code{\link[terra:extract]{terra::extract()}}.}
}
\value{
A data frame.
}
\description{
Polygon summaries of GEDI Level 4B metrics
}
