% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeling.R
\name{predictGEDI}
\alias{predictGEDI}
\title{Predict GEDI footprint values}
\usage{
predictGEDI(model, data, name = "prediction")
}
\arguments{
\item{model}{A \code{gedi_model} or fitted R model.}

\item{data}{Predictor table.}

\item{name}{Output column name.}
}
\value{
A data table containing input data and predictions.
}
\description{
Predict GEDI footprint values
}
