% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeling.R
\name{predictGEDIH5}
\alias{predictGEDIH5}
\title{Predict values and write them to HDF5}
\usage{
predictGEDIH5(model, data, output, dataset = "prediction")
}
\arguments{
\item{model}{A \code{gedi_model} or fitted R model.}

\item{data}{Predictor table.}

\item{output}{Output HDF5 filename.}

\item{dataset}{Dataset name.}
}
\value{
Invisibly returns \code{output}.
}
\description{
Predict values and write them to HDF5
}
