% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rasterProducts.R
\name{projectLevel4B}
\alias{projectLevel4B}
\title{Project a GEDI Level 4B raster}
\usage{
projectLevel4B(x, crs, method = "bilinear", ...)
}
\arguments{
\item{x}{A raster path or \code{SpatRaster}.}

\item{crs}{Target CRS.}

\item{method}{Resampling method.}

\item{...}{Passed to \code{\link[terra:project]{terra::project()}}.}
}
\value{
A \code{\link[terra:SpatRaster-class]{terra::SpatRaster}}.
}
\description{
Project a GEDI Level 4B raster
}
