% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling.R
\name{randomSampling}
\alias{randomSampling}
\title{GEDI random sampling specification}
\usage{
randomSampling(size)
}
\arguments{
\item{size}{Number or fraction of observations to select.}
}
\description{
GEDI random sampling specification
}
