% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeling.R
\name{rasterizeGEDI}
\alias{rasterizeGEDI}
\title{Rasterize GEDI footprint data}
\usage{
rasterizeGEDI(x, metric, res = 0.01, fun = mean, lon = NULL, lat = NULL, ...)
}
\arguments{
\item{x}{GEDI footprint table.}

\item{metric}{Metric column.}

\item{res}{Resolution in decimal degrees.}

\item{fun}{Aggregation function.}

\item{lon, lat}{Optional coordinate columns.}

\item{...}{Passed to \code{\link[terra:rasterize]{terra::rasterize()}}.}
}
\value{
A \code{\link[terra:SpatRaster-class]{terra::SpatRaster}}.
}
\description{
Rasterize GEDI footprint data
}
