% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/level4A.R
\name{rasterizeLevel4A}
\alias{rasterizeLevel4A}
\title{Rasterize GEDI Level 4A footprints}
\usage{
rasterizeLevel4A(
  level4A,
  metric = "agbd",
  fun = mean,
  res = 0.01,
  filename = "",
  overwrite = FALSE,
  ...
)
}
\arguments{
\item{level4A}{A table returned by \code{\link[=getLevel4A]{getLevel4A()}}.}

\item{metric}{Metric column name.}

\item{fun}{Aggregation function.}

\item{res}{Output resolution in decimal degrees.}

\item{filename}{Optional output GeoTIFF path.}

\item{overwrite}{Logical; overwrite \code{filename} when it exists.}

\item{...}{Additional arguments passed to \code{fun}.}
}
\value{
A \code{\link[terra:SpatRaster-class]{terra::SpatRaster}}.
}
\description{
Rasterize GEDI Level 4A footprints
}
