% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling.R
\name{rasterSampling}
\alias{rasterSampling}
\title{GEDI raster-stratified sampling specification}
\usage{
rasterSampling(size, raster, lon = NULL, lat = NULL)
}
\arguments{
\item{size}{Number or fraction per raster class.}

\item{raster}{A categorical \code{SpatRaster}.}

\item{lon, lat}{Optional coordinate column names.}
}
\description{
GEDI raster-stratified sampling specification
}
