% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rasterProducts.R
\name{readLevel3}
\alias{readLevel3}
\title{Read GEDI Level 3 gridded metrics}
\usage{
readLevel3(x, metrics = NULL)
}
\arguments{
\item{x}{One or more local paths, GDAL-readable URLs, or a \code{SpatRaster}.}

\item{metrics}{Optional layer names to retain.}
}
\value{
A \code{\link[terra:SpatRaster-class]{terra::SpatRaster}}.
}
\description{
Read GEDI Level 3 gridded metrics
}
\seealso{
\url{https://daac.ornl.gov/GEDI/guides/GEDI_L3_LandSurface_Metrics_V3.html}
}
