% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/level4A.R
\name{readLevel4A}
\alias{readLevel4A}
\title{Read a GEDI Level 4A granule}
\usage{
readLevel4A(level4Apath)
}
\arguments{
\item{level4Apath}{Path to a GEDI04_A HDF5 file.}
}
\value{
A \code{\link[=gedi.level4a-class]{gedi.level4a}} object. Close it with \code{\link[=close]{close()}}.
}
\description{
Opens a GEDI04_A HDF5 granule containing footprint-level aboveground
biomass density estimates.
}
\seealso{
\url{https://daac.ornl.gov/GEDI/guides/GEDI_L4A_AGB_Density_V3.html}
}
