% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rasterProducts.R
\name{readLevel4B}
\alias{readLevel4B}
\title{Read GEDI Level 4B gridded biomass}
\usage{
readLevel4B(x, metrics = NULL)
}
\arguments{
\item{x}{One or more local paths, GDAL-readable URLs, or a \code{SpatRaster}.}

\item{metrics}{Optional layer names to retain.}
}
\value{
A \code{\link[terra:SpatRaster-class]{terra::SpatRaster}}.
}
\description{
Read GEDI Level 4B gridded biomass
}
\seealso{
\doi{10.3334/ORNLDAAC/2299}
}
