% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rasterProducts.R
\name{resampleLevel3}
\alias{resampleLevel3}
\title{Resample a GEDI Level 3 raster}
\usage{
resampleLevel3(x, template, method = "bilinear", ...)
}
\arguments{
\item{x}{A raster path or \code{SpatRaster}.}

\item{template}{Target \code{SpatRaster}.}

\item{method}{Resampling method.}

\item{...}{Passed to \code{\link[terra:resample]{terra::resample()}}.}
}
\value{
A \code{\link[terra:SpatRaster-class]{terra::SpatRaster}}.
}
\description{
Resample a GEDI Level 3 raster
}
