% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rGEDI-package.R
\name{rGEDI-package}
\alias{rGEDI-package}
\title{rGEDI: An R Package for NASA's Global Ecosystem Dynamics Investigation (GEDI) Data Visualizing and Processing.}
\description{
The rGEDI package provides functions for i) downloading, ii) visualizing, iii) clipping, iv) gridding, iv) simulating and v) exporting GEDI data.
}
\note{
See more details about GEDI data in \url{https://gedi.umd.edu/dataproducts/products/}.
}
\seealso{
For comprehensive examples refer to \url{https://github.com/carlos-alberto-silva/rGEDI/blob/master/README.md}
}
\author{
Carlos A. Silva, Caio Hamamura, Ruben Valbuena, Steve Hancock, Adrian Cardil, Eben N. Broadbent, Danilo R. A. de Almeida, Celso H. L. Silva Junior and Carine Klauberg
}
