% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling.R
\name{sampleGEDI}
\alias{sampleGEDI}
\title{Sample GEDI footprint observations}
\usage{
sampleGEDI(x, method)
}
\arguments{
\item{x}{A data frame or data table of GEDI footprints.}

\item{method}{A specification returned by one of the sampling constructors.}
}
\value{
A \link[data.table:data.table]{data.table::data.table}.
}
\description{
Sample GEDI footprint observations
}
