% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling.R
\name{spacedSampling}
\alias{spacedSampling}
\title{GEDI minimum-distance sampling specification}
\usage{
spacedSampling(size, radius, lon = NULL, lat = NULL)
}
\arguments{
\item{size}{Number or fraction of observations to select.}

\item{radius}{Minimum separation in meters.}

\item{lon, lat}{Optional coordinate column names.}
}
\description{
GEDI minimum-distance sampling specification
}
