% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling.R
\name{stratifiedSampling}
\alias{stratifiedSampling}
\title{GEDI attribute-stratified sampling specification}
\usage{
stratifiedSampling(size, variable, breaks = NULL)
}
\arguments{
\item{size}{Number or fraction per stratum.}

\item{variable}{Column used as strata, or a numeric column to bin.}

\item{breaks}{Optional breaks for a numeric variable.}
}
\description{
GEDI attribute-stratified sampling specification
}
