% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdalBindings.R
\name{*.GDALRasterBand}
\alias{*.GDALRasterBand}
\title{GDALRasterBand * operation}
\usage{
\method{*}{GDALRasterBand}(first, second)
}
\arguments{
\item{first}{GDALRasterBand. The raster band to process.}

\item{second}{Atomic or GDALRasterBand. The other object to operate.}
}
\value{
An object of GDALRasterBand R6 class.
}
\description{
This function gives access to the GDALRasterBand using [[i]], where i is the band index to return.
}
