% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling.R
\name{to_vect}
\alias{to_vect}
\title{Convert GEDI footprints to a spatial vector}
\usage{
to_vect(x, lon = NULL, lat = NULL, crs = "EPSG:4326")
}
\arguments{
\item{x}{GEDI footprint table.}

\item{lon, lat}{Optional coordinate column names.}

\item{crs}{Coordinate reference system.}
}
\value{
A \code{\link[terra:SpatVector-class]{terra::SpatVector}}.
}
\description{
Convert GEDI footprints to a spatial vector
}
