% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gee.R
\name{vect_as_ee}
\alias{vect_as_ee}
\title{Convert a vector object to an Earth Engine feature collection}
\usage{
vect_as_ee(x)
}
\arguments{
\item{x}{\code{sf}, \code{sfc}, \code{SpatVector}, or vector file path.}
}
\value{
An Earth Engine \code{FeatureCollection}.
}
\description{
Convert a vector object to an Earth Engine feature collection
}
