% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clipLevel1B.R
\name{clipLevel1B}
\alias{clipLevel1B}
\title{Clip GEDI Level1B data by Coordinates}
\usage{
clipLevel1B(level1b, xleft, xright, ybottom, ytop, output)
}
\arguments{
\item{level1b}{A GEDI Level1B object (output of \code{\link[rGEDI:readLevel1B]{readLevel1B}} function). A S4 object of class "gedi.level1b".}

\item{xleft}{Numeric. West longitude (x) coordinate of bounding rectangle, in decimal degrees.}

\item{xright}{Numeric. East longitude (x) coordinate of bounding rectangle, in decimal degrees.}

\item{ybottom}{Numeric. South latitude (y) coordinate of bounding rectangle, in decimal degrees.}

\item{output}{Optional character path where to save the new hdf5file. The default stores a temporary file only.}

\item{ytopNumeric.}{North latitude (y) coordinate of bounding rectangle, in decimal degrees.}
}
\value{
An S4 object of class "gedi.level1b".
}
\description{
This function clips GEDI Level1B data within given bounding coordinates
}
\examples{
# specify the path to GEDI data
level1bpath <- system.file("extdata", "GEDIexample_level01B.h5", package="rGEDI")

# Reading GEDI level1B file
level1b<-readLevel1b(level1bpath)

# Bounding rectangle coordinates
xleft = -116.4683
xright = -116.5583
ybottom = 46.75208
ytop = 46.84229

# clip by extent boundary box
level1b_clip <- clipLevel1B(level1b,xleft, xright, ybottom, ytop)

}
