% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clipLevel1Bh5.R
\name{clipLevel1B}
\alias{clipLevel1B}
\title{Clip GEDI Level1B data}
\usage{
clipLevel1B(level1b, xleft, xright, ybottom, ytop, output = "")
}
\arguments{
\item{level1b}{h5file; S4 object of class H5File}

\item{xleft}{numeric. left x coordinates of rectangles.}

\item{xright}{numeric. right x coordinates of rectangles.}

\item{ybottom}{numeric. bottom y coordinates of rectangles.}

\item{ytop}{numeric. top y coordinates of rectangles.}

\item{output}{optional character path where to save the new h5file. Default "" (temporary file).}
}
\value{
Returns An object of class H5File; subset of LVIS Level1 data
}
\description{
Clip LVIS Level1 data within a given bounding coordinates
}
\examples{

#' LVIS level 2 file path
#level1bpath = system.file("extdata", "lvis_level1_clip.h5", package="rLVIS")

# Rectangle
#xleft = 82
#xright = 82.8
#ybottom = 0
#ytop = 20

#' Reading LVIS level 2 file
#level1_waveform = readLevel1B(level1bpath)

#output = tempfile(fileext=".h5")

#clipped_waveform = clipLevel1B(level1_waveform, output, xleft, xright, ybottom, ytop)

}
\author{
Caio Hamamura
}
