% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clipLevel1BGeo.R
\name{clipLevel1BGeo}
\alias{clipLevel1BGeo}
\title{Clip Level1BGeo data by Coordinates}
\usage{
clipLevel1BGeo(level1BGeo, xleft, xright, ybottom, ytop)
}
\arguments{
\item{level1BGeo}{A GEDI Level1b object (output of \code{\link[rGEDI:readLevel1B]{readLevel1B}} function). A S4 object of class "gedi.level1b".}

\item{xleft}{Numeric. West longitude (x) coordinate of bounding rectangle, in decimal degrees.}

\item{xright}{Numeric. East longitude (x) coordinate of bounding rectangle, in decimal degrees.}

\item{ybottom}{Numeric. South latitude (y) coordinate of bounding rectangle, in decimal degrees.}

\item{ytop}{Numeric. North latitude (y) coordinate of bounding rectangle, in decimal degrees.}
}
\value{
An S4 object of class \code{\link[data.table:data.table]{data.table-class}}.
}
\description{
This function clips GEDI level1B extracted geolocation (level1BGeo)
data within given bounding coordinates
}
\examples{
# specify the path to GEDI Level 1B data
level1bpath <- system.file("extdata", "GEDIexample_level01B.h5", package="rGEDI")

# Reading GEDI level1B data
level1b <- readLevel1B(level1bpath)

# Get GEDI level1B geolocations
level1BGeo<-getLevel1BGeo(level1b)

# Bounding rectangle coordinates
xleft = -116.4683
xright = -116.5583
ybottom = 46.75208
ytop = 46.84229

# clip by boundary box coordinates
level1bGeo_clip <- clipLevel1BGeo(level1BGeo,xleft, xright, ybottom, ytop)

library(leaflet)
leaflet() \%>\%
 addCircleMarkers(level1bGeo_clip@dt$longitude_bin0,
                  level1bGeo_clip@dt$latitude_bin0,
                  radius = 1,
                  opacity = 1,
                  color = "red")  \%>\%
 addScaleBar(options = list(imperial = FALSE)) \%>\%
 addProviderTiles(providers$Esri.WorldImagery)
}
\seealso{
https://lpdaac.usgs.gov/products/gedi01_bv001/
}
