% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clipLevel1BGeo.R
\name{clipLevel1BGeo}
\alias{clipLevel1BGeo}
\title{Clip Level1BGeo data by Coordinates}
\usage{
clipLevel1BGeo(level1BGeo, xmin, xmax, ymin, ymax)
}
\arguments{
\item{level1BGeo}{A GEDI Level1b object (output of \code{\link[rGEDI:readLevel1B]{readLevel1B}} function). A S4 object of class "gedi.level1b".}

\item{xmin}{Numeric. West longitude (x) coordinate of bounding rectangle, in decimal degrees.}

\item{xmax}{Numeric. East longitude (x) coordinate of bounding rectangle, in decimal degrees.}

\item{ymin}{Numeric. South latitude (y) coordinate of bounding rectangle, in decimal degrees.}

\item{ymax}{Numeric. North latitude (y) coordinate of bounding rectangle, in decimal degrees.}
}
\value{
An S4 object of class \code{\link[data.table:data.table]{data.table-class}}.
}
\description{
This function clips GEDI level1B extracted geolocation (level1BGeo)
data within given bounding coordinates
}
\examples{
\dontrun{
# specify the path to download GEDI example dataset
outdir<-getwd()

# downloading GEDI example dataset (zip file)
download.file(
             paste0(
                    "https://github.com/carlos-alberto-silva/rGEDI/",
                    "releases/download/examples/examples.zip"
             ),
             destfile=paste0(outdir,"/examples.zip"))

# unzip the file
unzip(paste0(outdir,"\\\\examples.zip"))

# specify the path to GEDI level1B data
level1bpath = paste0(outdir,"\\\\GEDI01_B_2019108080338_O01964_T05337_02_003_01_sub.h5")

# Reading GEDI level1B file
level1b<-readLevel1b(gedilevel1b)

# Bounding rectangle coordinates
xmin = -44.15036
xmax = -44.10066
ymin = -13.75831
ymax = -13.71244

# clip by boundary box coordinates
level1bGeo_clip <- clipLevel1BGeo(level1BGeo,xmin, xmax, ymin, ymax)

library(leaflet)
leaflet() \%>\%
 addCircleMarkers(level1bGeo_clip@dt$longitude_bin0,
                  level1bGeo_clip@dt$latitude_bin0,
                  radius = 1,
                  opacity = 1,
                  color = "red")  \%>\%
 addScaleBar(options = list(imperial = FALSE)) \%>\%
 addProviderTiles(providers$Esri.WorldImagery)
}
}
\seealso{
https://lpdaac.usgs.gov/products/gedi01_bv001/
}
