% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clipLevel1BGeo.R
\name{clipLevel1BGeoGeometry}
\alias{clipLevel1BGeoGeometry}
\title{Clip GEDI Full Waveform Geolocations by geometry}
\usage{
clipLevel1BGeoGeometry(level1BGeo, polygon, split_by = "id")
}
\arguments{
\item{level1BGeo}{A \code{\link[data.table:data.table-class]{data.table::data.table}} resulting from \code{\link[=getLevel1BGeo]{getLevel1BGeo()}} function.}

\item{polygon}{Polygon. An object of class \code{\link[sf:sf]{sf::sf}},
which can be loaded as an ESRI shapefile using \link[sf:st_read]{sf::st_read} function in the \emph{sf} package.}

\item{split_by}{Polygon id. If defined, GEDI data will be clipped by each polygon using the
polygon id from table of attribute defined by the user.}
}
\value{
Returns an S4 object of class \code{\link[data.table:data.table-class]{data.table::data.table}} containing the
clipped GEDI level1B extracted geolocations.
}
\description{
This function clips level1BGeo extracted geolocation (level1BGeo)
data within a given geometry
}
\examples{
# Specifying the path to GEDI level1B data (zip file)
outdir <- tempdir()
level1B_fp_zip <- system.file("extdata",
  "GEDI01_B_2019108080338_O01964_T05337_02_003_01_sub.zip",
  package = "rGEDI"
)

# Unzipping GEDI level1B data
level1Bpath <- unzip(level1B_fp_zip, exdir = outdir)

# Reading GEDI level1B data (h5 file)
level1b <- readLevel1B(level1Bpath = level1Bpath)

# Extracting GEDI Full Waveform Geolocations
level1BGeo <- getLevel1BGeo(level1b)

# Specifying the path to shapefile
polygon_filepath <- system.file("extdata", "stands_cerrado.shp", package = "rGEDI")

# Reading shapefile as sf object
library(sf)
polygon <- sf::st_read(polygon_filepath)

# Clipping GEDI Full Waveform Geolocations by Geometry
level1BGeo_clip <- clipLevel1BGeoGeometry(level1BGeo, polygon, split_by = "id")

if (interactive() && requireNamespace("leaflet", quietly = TRUE)) {
  library(leaflet)
  leaflet() \%>\%
    addCircleMarkers(level1BGeo_clip$longitude_bin0,
      level1BGeo_clip$latitude_bin0,
      radius = 1,
      opacity = 1,
      color = "red"
    ) \%>\%
    addScaleBar(options = list(imperial = FALSE)) \%>\%
    addPolygons(
      data = polygon, weight = 1, col = "white",
      opacity = 1, fillOpacity = 0
    ) \%>\%
    addProviderTiles(providers$Esri.WorldImagery)
}

close(level1b)
}
\seealso{
\url{https://www.earthdata.nasa.gov/data/catalog/lpcloud-gedi01-b-002}
}
