% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clipLevel1Bh5.R
\name{clipLevel1Bgeometry}
\alias{clipLevel1Bgeometry}
\title{Clip LVIS Level1 data by geometry}
\usage{
clipLevel1Bgeometry(level1b, polygon_spdf, output = "")
}
\arguments{
\item{level1b}{h5file; S4 object of class H5File}

\item{polygon_spdf}{SpatialDataFrame. A polygon dataset for clipping the waveform}

\item{output}{optional character path where to save the new h5file. Default "" (temporary file).}
}
\value{
Returns An object of class H5File; subset of LVIS Level1 data
}
\description{
Clip LVIS Level1 data within a given bounding coordinates
}
\examples{

#' LVIS level 2 file path
#level1_filepath = system.file("extdata", "lvis_level1_clip.h5", package="rLVIS")

#' Reading LVIS level 2 file
#level1_waveform = readLevel1b(level1_filepath)

# Polgons file path
#polygons_filepath <- system.file("extdata", "LVIS_Mondah_clip_polygon.shp", package="rLVIS")

# Reading LVIS level 2 file
#polygon_spdf<-raster::shapefile(polygons_filepath)

#output = tempfile(fileext="h5")

#clipped_waveform = clipLevel1Geometry(level1_waveform, output, polygon_spdf)

}
\author{
Caio Hamamura
}
