% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clipLevel1B.R, R/clipLevel1BGeo.R
\name{clipLevel1BGeometry}
\alias{clipLevel1BGeometry}
\title{Clip GEDI Level1B data by geometry}
\usage{
clipLevel1BGeometry(level1b, polygon_spdf, output)

clipLevel1BGeo(level1BGeo, polygon_spdf, output)
}
\arguments{
\item{level1BGeo}{A GEDI Level1b object (output of \code{\link[rGEDI:readLevel1B]{readLevel1B}} function). A S4 object of class "gedi.level1b".}

\item{polygon_spdf}{Polygon. An object of class \code{\link[sp]{SpatialPolygonsDataFrame-class}},
which can be loaded as an ESRI shapefile using \code{\link[rgdal:readOGR]{readOGR}} function in the \emph{rgdal} package.}

\item{split_by}{Polygon id. If defined, GEDI data will be clipped by each polygon using the polygon id from table of attribute defined by the user}

\item{level1b}{A GEDI Level1B object (output of \code{\link[rGEDI:readLevel1B]{readLevel1B}} function). A S4 object of class "gedi.level1b".}

\item{output}{Optional character path where to save the new hdf5file. The default stores a temporary file only.}

\item{polygon_spdf}{Polygon. An object of class \code{\link[sp]{SpatialPolygonsDataFrame-class}},
which can be loaded as an ESRI shapefile using \code{\link[rgdal:readOGR]{readOGR}} function in the \emph{rgdal} package.}
}
\value{
An S4 object of class "gedi.level1b".

A S4 object of class \code{\link[data.table:data.table]{data.table-class}}.
}
\description{
This function clips GEDI Level1B data within a bounding geometry

This function clips GEDI level1B extracted geolocation (level1BGeo)
data within given geometry
}
\examples{
# specify the path to GEDI data
level1bpath <- system.file("extdata", "GEDIexample_level01B.h5", package="rGEDI")

# Reading GEDI level1B file
level1b<-readLevel1b(level1bpath)

# specify the path to shapefile
polygon_filepath <- system.file("extdata", "clip_polygon.shp", package="rGEDI")

# Reading shapefile as SpatialPolygonsDataFrame object
library(rgdal)
polygon_spdf<-readOGR(polygons_filepath)

# clip by extent boundary box
level1b_clip <- clipLevel1BGeometry(level1b, polygon_spdf = polygon_spdf)


# specify the path to GEDI Level 1B data
level1bpath <- system.file("extdata", "GEDIexample_level01B.h5", package="rGEDI")

# Reading GEDI level1B data
level1b <- readLevel1B(level1bpath)

# Get GEDI level1B geolocations
level1BGeo<-getLevel1BGeo(level1b)

# specify the path to shapefile
polygon_filepath <- system.file("extdata", "clip_polygon.shp", package="rGEDI")

# Reading shapefile as SpatialPolygonsDataFrame object
library(rgdal)
polygon_spdf<-readOGR(polygons_filepath)

level1bGeo_clip = clipLevel1BGeometry(level1bGeo, polygon_spdf, split_by="id")

library(leaflet)
leaflet() \%>\%
 addCircleMarkers(level1b_clip@dt$longitude_bin0,
                  level1b_clip@dt$latitude_bin0,
                  radius = 1,
                  opacity = 1,
                  color = "red")  \%>\%
 addScaleBar(options = list(imperial = FALSE)) \%>\%
 addPolygons(data=polygon_spdf,weight=1,col = 'white',
             opacity = 1, fillOpacity = 0) \%>\%
 addProviderTiles(providers$Esri.WorldImagery)
}
\seealso{
https://lpdaac.usgs.gov/products/gedi01_bv001/
}
