% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clipLevel2A.R
\name{clipLevel2A}
\alias{clipLevel2A}
\title{Clip GEDI Level2A data by Coordinates}
\usage{
clipLevel2A(level2a, xmin, xmax, ymin, ymax, output)
}
\arguments{
\item{level2a}{A GEDI Level2A object (output of \code{\link[rGEDI:readLevel2A]{readLevel2A}} function). A S4 object of class "gedi.level2a".}

\item{xmin}{Numeric. West longitude (x) coordinate of bounding rectangle, in decimal degrees.}

\item{xmax}{Numeric. East longitude (x) coordinate of bounding rectangle, in decimal degrees.}

\item{ymin}{Numeric. South latitude (y) coordinate of bounding rectangle, in decimal degrees.}

\item{ymax}{Numeric. North latitude (y) coordinate of bounding rectangle, in decimal degrees.}

\item{output}{Optional character path where to save the new hdf5file. The default stores a temporary file only.}
}
\value{
Returns a list of S4 object of class "gedi.level2a".
}
\description{
This function clips GEDI Level2A data within given bounding coordinates
}
\examples{
# specify the path to GEDI level2A data (zip file)
level2A_fp_zip <- system.file("extdata",
                  "GEDI02_A_2019108080338_O01964_T05337_02_001_01_sub.zip",
                  package="rGEDI")

# Unzipping GEDI level2A data
level2Apath <- unzip(level2A_fp_zip,exdir = dirname(level2A_fp_zip))

# Reading GEDI level2A data (h5 file)
level2a<-readLevel2A(level2Apath=level2Apath)

# Bounding rectangle coordinates
xmin = -44.15036
xmax = -44.10066
ymin = -13.75831
ymax = -13.71244

# Spepecifing output file and path
output<-paste0(getwd(),"//GEDI02_A_2019108080338_O01964_T05337_02_001_01_clip")

# clip by extent boundary box
level2a_clip <- clipLevel2A(level1a,xmin,xmax,ymin,ymax,output)

}
\seealso{
https://lpdaac.usgs.gov/products/gedi02_av001/
}
