% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clipLevel2A.R
\name{clipLevel2AGeometry}
\alias{clipLevel2AGeometry}
\title{Clip GEDI Level2A data by geometry}
\usage{
clipLevel2AGeometry(level2a, polygon_spdf, output="", split_by=NULL)
}
\arguments{
\item{level2a}{A GEDI Level2A object (output of \code{\link[rGEDI:readLevel2A]{readLevel2A}} function). A S4 object of class "gedi.level2a".}

\item{polygon_spdf}{Polygon. An object of class \code{\link[sp]{SpatialPolygonsDataFrame-class}},
which can be loaded as an ESRI shapefile using \code{\link[raster:shapefile]{raster::shapefile()}} function in the \emph{raster} package.}

\item{output}{optional character path where to save the new h5file. Default "" (temporary file).}

\item{split_by}{Polygon id. If defined, GEDI data will be clipped by each polygon using the polygon id from table of attribute defined by the user}
}
\value{
Returns a S4 object of class "gedi.level2a".
}
\description{
This function clips GEDI Level2A data within given geometry
}
\examples{
\dontrun{
# specify the path to download GEDI example dataset
outdir<-getwd()

# downloading GEDI example dataset (zip file)
download.file(
             paste0(
                    "https://github.com/carlos-alberto-silva/rGEDI/",
                    "releases/download/examples/examples.zip"
             ),
             destfile=paste0(outdir,"/examples.zip"))

# specify the path to GEDI level2A data
level2apath = paste0(outdir,"\\\\GEDI02_A_2019108080338_O01964_T05337_02_001_01_sub.h5")

# Reading GEDI level2A data
level2a<-readLevel2A(level2apath)

# specify the path to shapefile
polygon_filepath <- system.file("extdata", "stands_cerrado.shp", package="rGEDI")

# Reading shapefile as SpatialPolygonsDataFrame object
library(raster)
polygon_spdf<-shapefile(polygons_filepath)

level2a_clip <- clipLevel2AGeometry(level2a, polygon_spdf)
}

}
\seealso{
https://lpdaac.usgs.gov/products/gedi02_av001/
}
