% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clipLevel2A.R
\name{cliplevel2ah5Geometry}
\alias{cliplevel2ah5Geometry}
\title{Clip GEDI Level2A data by geometry}
\usage{
clipLevel2a(level2a, polygon_spdf, output)
}
\arguments{
\item{level2a}{A GEDI Level2A object (output of \code{\link[rGEDI:readLevel2A]{readLevel2A}} function). A S4 object of class "gedi.level2a".}

\item{polygon_spdf}{Polygon. An object of class \code{\link[sp]{SpatialPolygonsDataFrame-class}},
which can be loaded as an ESRI shapefile using \code{\link[rgdal:readOGR]{readOGR}} function in the \emph{rgdal} package.}

\item{output}{optional character path where to save the new h5file. Default "" (temporary file).}
}
\value{
Returns a S4 object of class "gedi.level2a".
}
\description{
This function clips GEDI Level2A data within given geometry
}
\examples{

# specify the path to GEDI level2A data
level2apath <- system.file("extdata", "GEDIexample_level02A.h5", package="rGEDI")

# Reading GEDI level2A data
level2a<-readLevel2a(level1bpath)

# specify the path to shapefile
polygon_filepath <- system.file("extdata", "clip_polygon.shp", package="rGEDI")

# Reading shapefile as SpatialPolygonsDataFrame object
library(rgdal)
polygon_spdf<-readOGR(polygons_filepath)

level2a_clip <- clipLevel2AGeometry(level2a,polygon_spdf)

}
