% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clipLevel2AM.R
\name{clipLevel2AM}
\alias{clipLevel2AM}
\title{Clip GEDI level2adt data}
\usage{
clipLevel2AM(x, xleft, xright, ybottom, ytop)
}
\arguments{
\item{level2adt}{level2adt; S4 object of class "gedi.level1b.dt"}

\item{extent}{Extent object of a Spatial object .}
}
\value{
Returns An object of class "gedi.level1b.dt"; subset of GEDI Level1B data
}
\description{
Clip GEDI level2adt data within a given bounding coordinates
}
\examples{

#' GEDI level1B file path
level1_filepath = system.file("extdata", "lvis_level1_clip.h5", package="rLVIS")

#' Reading LVIS level 2 file
level1_waveform = readLevel1b(level1Bfilepath)

# Rectangle area for cliping
xmin = -116.4683
xmax = -116.3177
ymin = 46.75208
ymax = 46.84229

# creating an exent object
ext<-extent(c(xmin=xmin, xmax=xmax, ymin=ymin, ymax=ymax))

clipped_level2adt = clipLevel1(level2adt, extent=ext)

library(leaflet)
leaflet() \%>\%
 addCircleMarkers(clipped_level2adt@dt$longitude_bin0,
                  clipped_level2adt@dt$latitude_bin0,
                  radius = 1,
                  opacity = 1,
                  color = "red")  \%>\%
 addScaleBar(options = list(imperial = FALSE)) \%>\%
 addProviderTiles(providers$Esri.WorldImagery)
}
