% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clipLevel2AM.R
\name{clipLevel2AM}
\alias{clipLevel2AM}
\title{Clip Level2AM data by Coordinates}
\usage{
cliplevel2AM(level2a, xleft, xright, ybottom, ytop, output)
}
\arguments{
\item{level2AM}{A GEDI Level2A object (output of \code{\link[rGEDI:readLevel2A]{readLevel2A}} function). A S4 object of class "gedi.level2a".}

\item{xleft}{Numeric. West longitude (x) coordinate of bounding rectangle, in decimal degrees.}

\item{xright}{Numeric. East longitude (x) coordinate of bounding rectangle, in decimal degrees.}

\item{ybottom}{Numeric. South latitude (y) coordinate of bounding rectangle, in decimal degrees.}

\item{ytop}{Numeric. North latitude (y) coordinate of bounding rectangle, in decimal degrees.}

\item{output}{Optional character path where to save the new hdf5file. The default stores a temporary file only.}
}
\value{
An S4 object of class "gedi.level2a".
}
\description{
This function clips GEDI Level2A extracted Elevation and Height Metrics (Level2AM)
within given bounding coordinates
}
\examples{
#' GEDI level2A file path
level2afilepath = system.file("extdata", "lvis_level1_clip.h5", package="rGEDI")

#' Reading GEDI level2A file
level2a = readLevel2A(level2afilepath)

#' Extracting GEDI Elevation and Height Metrics
level2AM = getLevel2AM(level2a)

# Bounding rectangle coordinates
xleft = -116.4683
xright = -116.5583
ybottom = 46.75208
ytop = 46.84229

# clip by extent boundary box
level2AM_clip <- clipLevel2AM(level2AM,xleft, xright, ybottom, ytop)

}
\seealso{
https://lpdaac.usgs.gov/products/gedi02_av001/
}
