% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clipLevel2B.R
\name{clipLevel2B}
\alias{clipLevel2B}
\title{Clip GEDI Level2B data by Coordinates}
\usage{
clipLevel2B(level2b, xmin, xmax, ymin, ymax, output = "")
}
\arguments{
\item{level2b}{A GEDI Level2B object (output of \code{\link[rGEDI:readLevel2B]{readLevel2B}} function). A S4 object of class "gedi.level2b".}

\item{xmin}{Numeric. West longitude (x) coordinate of bounding rectangle, in decimal degrees.}

\item{xmax}{Numeric. East longitude (x) coordinate of bounding rectangle, in decimal degrees.}

\item{ymin}{Numeric. South latitude (y) coordinate of bounding rectangle, in decimal degrees.}

\item{ymax}{Numeric. North latitude (y) coordinate of bounding rectangle, in decimal degrees.}

\item{output}{Optional character path where to save the new hdf5 file. The default stores a temporary file only.}
}
\value{
An S4 object of class "gedi.level2b".
}
\description{
This function extracts GEDI Level1B data within given bounding coordinates
}
\examples{
\dontrun{
# specify the path to download GEDI example dataset
outdir<-getwd()

# downloading GEDI example dataset (zip file)
download.file(
             paste0(
                    "https://github.com/carlos-alberto-silva/rGEDI/",
                    "releases/download/examples/examples.zip"
             ),
             destfile=paste0(outdir,"/examples.zip"))

# unzip the file
unzip(paste0(outdir,"\\\\examples.zip"))

# specify the path to GEDI level2B data
level2bpath = paste0(outdir,"\\\\GEDI02_B_2019108080338_O01964_T05337_02_001_01_sub.h5")

# Reading GEDI level1B file
level2b<-readLevel2b(gedilevel2b)

# Bounding rectangle coordinates
xmin = -44.15036
xmax = -44.10066
ymin = -13.75831
ymax = -13.71244

# clip level2BVPM by extent boundary box
level2b_clip <- level2BVPM(level2BVPM,xmin, xmax, ymin, ymax)

library(leaflet)
leaflet() \%>\%
 addCircleMarkers(level2b_clip@dt$longitude_bin0,
                  level2b_clip@dt$latitude_bin0,
                  radius = 1,
                  opacity = 1,
                  color = "red")  \%>\%
 addScaleBar(options = list(imperial = FALSE)) \%>\%
 addPolygons(data=polygon_spdf,weight=1,col = 'white',
             opacity = 1, fillOpacity = 0) \%>\%
 addProviderTiles(providers$Esri.WorldImagery)

}
}
\seealso{
https://lpdaac.usgs.gov/products/gedi01_bv001/
}
