% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clipLevel2B.R
\name{clipLevel2B}
\alias{clipLevel2B}
\title{Clip GEDI Level2B data by Coordinates}
\usage{
clipLevel2B(level2b, xmin, xmax, ymin, ymax, output = "")
}
\arguments{
\item{level2b}{A GEDI Level2B object (output of \code{\link[rGEDI:readLevel2B]{readLevel2B}} function). A S4 object of class "gedi.level2b".}

\item{xmin}{Numeric. West longitude (x) coordinate of bounding rectangle, in decimal degrees.}

\item{xmax}{Numeric. East longitude (x) coordinate of bounding rectangle, in decimal degrees.}

\item{ymin}{Numeric. South latitude (y) coordinate of bounding rectangle, in decimal degrees.}

\item{ymax}{Numeric. North latitude (y) coordinate of bounding rectangle, in decimal degrees.}

\item{output}{Optional character path where to save the new hdf5 file. The default stores a temporary file only.}
}
\value{
Returns a list of S4 object of class "gedi.level2b".
}
\description{
This function extracts GEDI Level1B data within given bounding coordinates
}
\examples{
# specify the path to GEDI level2B data (zip file)
level2B_fp_zip <- system.file("extdata",
                  "GEDI02_B_2019108080338_O01964_T05337_02_001_01_sub.zip",
                  package="rGEDI")

# Unzipping GEDI level2A data
level2Bpath <- unzip(level2B_fp_zip,exdir = dirname(level2B_fp_zip))

# Reading GEDI level2B data (h5 file)
level2b<-readLevel2B(level2Bpath=level2Bpath)

# Bounding rectangle coordinates
xmin = -44.15036
xmax = -44.10066
ymin = -13.75831
ymax = -13.71244

# Spepecifing output file and path
output<-paste0(getwd(),"//GEDI02_B_2019108080338_O01964_T05337_02_001_01_clip")

# clip level2BVPM by extent boundary box
level2b_clip <- level2BVPM(level2BVPM,xmin, xmax, ymin, ymax)

library(leaflet)
leaflet() \%>\%
 addCircleMarkers(level2b_clip@dt$longitude_bin0,
                  level2b_clip@dt$latitude_bin0,
                  radius = 1,
                  opacity = 1,
                  color = "red")  \%>\%
 addScaleBar(options = list(imperial = FALSE)) \%>\%
 addPolygons(data=polygon_spdf,weight=1,col = 'white',
             opacity = 1, fillOpacity = 0) \%>\%
 addProviderTiles(providers$Esri.WorldImagery)


}
\seealso{
https://lpdaac.usgs.gov/products/gedi01_bv001/
}
