% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clipLevel2B.R
\name{clipLevel2B}
\alias{clipLevel2B}
\title{Clip GEDI Level2B data by Coordinates}
\usage{
clipLevel2B(level2b, xleft, xright, ybottom, ytop, output="")
}
\arguments{
\item{xleft}{Numeric. West longitude (x) coordinate of bounding rectangle, in decimal degrees.}

\item{xright}{Numeric. East longitude (x) coordinate of bounding rectangle, in decimal degrees.}

\item{ybottom}{Numeric. South latitude (y) coordinate of bounding rectangle, in decimal degrees.}

\item{output}{Optional character path where to save the new hdf5 file. The default stores a temporary file only.}

\item{level1b}{A GEDI Level2B object (output of \code{\link[rGEDI:readLevel2B]{readLevel2B}} function). A S4 object of class "gedi.level2b".}

\item{ytopNumeric.}{North latitude (y) coordinate of bounding rectangle, in decimal degrees.}
}
\value{
An S4 object of class "gedi.level2b".
}
\description{
This function extracts GEDI Level1B data within given bounding coordinates
}
\examples{
# specify the path and data file and read it
level2bpath <- system.file("extdata", "GEDIexample_level02B.h5", package="rGEDI")

# reading GEDI level2B data
level2b <- readLevel2B(level2bpath)

# Bounding rectangle coordinates
xleft = -116.4683
xright = -116.5583
ybottom = 46.75208
ytop = 46.84229

# clip level2BVPM by extent boundary box
level2b_clip <- level2BVPM(level2BVPM,xleft, xright, ybottom, ytop)

library(leaflet)
leaflet() \%>\%
 addCircleMarkers(level2b_clip@dt$longitude_bin0,
                  level2b_clip@dt$latitude_bin0,
                  radius = 1,
                  opacity = 1,
                  color = "red")  \%>\%
 addScaleBar(options = list(imperial = FALSE)) \%>\%
 addPolygons(data=polygon_spdf,weight=1,col = 'white',
             opacity = 1, fillOpacity = 0) \%>\%
 addProviderTiles(providers$Esri.WorldImagery)
}
