% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clipLevel2BPAIProfile.R
\name{clipLevel2BPAIProfile}
\alias{clipLevel2BPAIProfile}
\title{Clip GEDI Plant Area Index profile data by Coordinates}
\usage{
clipLevel2BPAIProfile(level2BPAIProfile, xmin, xmax, ymin, ymax)
}
\arguments{
\item{level2BPAIProfile}{A GEDI Level2B object (output of \code{\link[rGEDI:getLevel2BPAIProfile]{getLevel2BPAIProfile}} function). A S4 object of class "gedi.level2b".}

\item{xmin}{Numeric. West longitude (x) coordinate of bounding rectangle, in decimal degrees.}

\item{xmax}{Numeric. East longitude (x) coordinate of bounding rectangle, in decimal degrees.}

\item{ymin}{Numeric. South latitude (y) coordinate of bounding rectangle, in decimal degrees.}

\item{ymax}{Numeric. North latitude (y) coordinate of bounding rectangle, in decimal degrees.}
}
\value{
An S4 object of class \code{\link[data.table:data.table]{data.table-class}}
containing the Plant Area Index profile data.
}
\description{
This function clips GEDI level2B-derived
Plant Area Index profile within given bounding coordinates
}
\examples{
\dontrun{
# specify the path to download GEDI example dataset
outdir<-getwd()

# downloading GEDI example dataset (zip file)
download.file(
             paste0(
                    "https://github.com/carlos-alberto-silva/rGEDI/",
                    "releases/download/examples/examples.zip"
             ),
             destfile=paste0(outdir,"/examples.zip"))

# unzip the file
unzip(paste0(outdir,"\\\\examples.zip"))

# specify the path to GEDI level2B data
level2bpath = paste0(outdir,"\\\\GEDI02_B_2019108080338_O01964_T05337_02_001_01_sub.h5")

# Reading GEDI level1B file
level2b<-readLevel2b(gedilevel2b)

# Get Plant Area Index profile
level2BPAIProfile<-getLevel2BPAIProfile(level2b)

# Bounding rectangle coordinates
xmin = -44.15036
xmax = -44.10066
ymin = -13.75831
ymax = -13.71244

# clip level2BVPM by extent boundary box
level2b_clip <- clipLevel2BPAIProfile(level2BPAIProfile,xmin, xmax, ymin, ymax)
}
}
\seealso{
https://lpdaac.usgs.gov/products/gedi02_bv001/
}
