% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clipLevel2BPAIProfile.R
\name{clipLevel2BPAIProfile}
\alias{clipLevel2BPAIProfile}
\title{Clip level2BPAIProfile data by Coordinates}
\usage{
clipLevel2BPAIProfile(level2BPAIProfile, xleft, xright, ybottom, ytop, output="")
}
\arguments{
\item{xleft}{Numeric. West longitude (x) coordinate of bounding rectangle, in decimal degrees.}

\item{xright}{Numeric. East longitude (x) coordinate of bounding rectangle, in decimal degrees.}

\item{ybottom}{Numeric. South latitude (y) coordinate of bounding rectangle, in decimal degrees.}

\item{level2BPAIProfile}{A GEDI Level2B object (output of \code{\link[rGEDI:getLevel2BPAIProfile]{getLevel2BPAIProfile}} function). A S4 object of class "gedi.level2b".}

\item{ytopNumeric.}{North latitude (y) coordinate of bounding rectangle, in decimal degrees.}

\item{output}{Optional character path where to save the new hdf5file. The default stores a temporary file only.}
}
\value{
An S4 object of class "gedi.level2b".
}
\description{
This function clips GEDI level2B-derived
Plant Area Index profile within given bounding coordinates
}
\examples{
# specify the path and data file and read it
level2bpath <- system.file("extdata", "GEDIexample_level02B.h5", package="rGEDI")

# Reading GEDI level2B data
level2b <- readLevel2B(level2bpath)

# Get Plant Area Index profile
level2BPAIProfile<-getLevel2BPAIProfile(level2b)

# Bounding rectangle coordinates
xleft = -116.4683
xright = -116.5583
ybottom = 46.75208
ytop = 46.84229

# clip level2BVPM by extent boundary box
level2b_clip <- clipLevel2BPAIProfile(level2BPAIProfile,xleft, xright, ybottom, ytop)

}
