% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clipLevel2BPAIProfile.R
\name{clipLevel2BPAIProfileGeometry}
\alias{clipLevel2BPAIProfileGeometry}
\title{Clip GEDI Plant Area Index profile data by geometry}
\usage{
clipLevel2BPAIProfileGeometry(level2BPAIProfile, polygon_spdf, split_by)
}
\arguments{
\item{level2BPAIProfile}{A GEDI Level2B object (output of \code{\link[rGEDI:getLevel2BPAIProfile]{getLevel2BPAIProfile}} function). A S4 object of class "data.table".}

\item{polygon_spdf}{Polygon. An object of class \code{\link[sp]{SpatialPolygonsDataFrame-class}},
which can be loaded as an ESRI shapefile using \code{\link[raster:shapefile]{raster::shapefile()}} function in the \emph{raster} package.}

\item{split_by}{Polygon id. If defined, GEDI data will be clipped by each polygon using the polygon id from table of attribute defined by the user}
}
\value{
An S4 object of class \code{\link[data.table:data.table]{data.table-class}}.
}
\description{
This function clips GEDI level2B-derived
Plant Area Index profile within given geometry
}
\examples{
\dontrun{
# specify the path to download GEDI example dataset
outdir<-getwd()

# downloading GEDI example dataset (zip file)
download.file(
             paste0(
                    "https://github.com/carlos-alberto-silva/rGEDI/",
                    "releases/download/examples/examples.zip"
             ),
             destfile=paste0(outdir,"/examples.zip"))

# unzip the file
unzip(paste0(outdir,"\\\\examples.zip"))

# specify the path to GEDI level2B data
level2bpath = paste0(outdir,"\\\\GEDI02_B_2019108080338_O01964_T05337_02_001_01_sub.h5")

# Reading GEDI level1B file
level2b<-readLevel2b(gedilevel2b)

# Get Plant Area Index profile
level2BPAIProfile<-getLevel2BPAIProfile(level2b)

# specify the path to shapefile
polygon_filepath <- system.file("extdata", "stands_cerrado.shp", package="rGEDI")

# Reading shapefile as SpatialPolygonsDataFrame object
library(raster)
polygon_spdf<-shapefile(polygons_filepath)

# clip level2BPAIProfile by geometry
level2b_clip_geometry <- clipLevel2BPAIProfileGeometry(
                                                      level2BPAIProfile,
                                                      polygon_spdf,
                                                      split_by="id")

}
}
\seealso{
https://lpdaac.usgs.gov/products/gedi02_bv001/
}
