<<<<<<< HEAD
% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clipLevel2BPAIProfile.R
\name{clipLevel2BPAIProfileGeometry}
\alias{clipLevel2BPAIProfileGeometry}
\title{Clip level2BPAIProfile data by geometry}
\usage{
clipLevel2BPAIProfileGeometry(level2BPAIProfile, polygon_spdf, output)
}
\arguments{
\item{polygon_spdf}{Polygon. An object of class \code{\link[sp]{SpatialPolygonsDataFrame-class}},
which can be loaded as an ESRI shapefile using \code{\link[rgdal:readOGR]{readOGR}} function in the \emph{rgdal} package.}

\item{split_by}{Polygon id. If defined, GEDI data will be clipped by each polygon using the polygon id from table of attribute defined by the user}

\item{level2BPAIProfile}{A GEDI Level2B object (output of \code{\link[rGEDI:getLevel2BPAIProfile]{getLevel2BPAIProfile}} function). A S4 object of class "gedi.level2b".}
}
\value{
An S4 object of class \code{\link[data.table:data.table]{data.table-class}}.
}
\description{
This function clips GEDI level2B-derived
Plant Area Index profile within given geometry
}
\examples{

# specify the path and data file and read it
level2bpath <- system.file("extdata", "GEDIexample_level02B.h5", package="rGEDI")

#'# reading GEDI level2B data
level2b <- readLevel2B(level2bpath)

# Get Plant Area Index profile
level2BPAIProfile<-getLevel2BPAIProfile(level2b)

# specify the path to shapefile
polygon_filepath <- system.file("extdata", "clip_polygon.shp", package="rGEDI")

# Reading shapefile as SpatialPolygonsDataFrame object
library(rgdal)
polygon_spdf<-readOGR(polygons_filepath)

# clip level2BPAIProfile by geometry
level2b_clip_geometry <- clipLevel2BPAIGeometry(level2BPAIProfile,polygon_spdf, split_by="id")

}
=======
% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clipLevel2BPAIProfile.R
\name{clipLevel2BPAIProfileGeometry}
\alias{clipLevel2BPAIProfileGeometry}
\title{Clip GEDI level2bdt data by geometry}
\usage{
clipLevel2BPAIProfileGeometry(x, polygon_spdf, split_by = NULL)
}
\arguments{
\item{polygon_spdf}{SpatialDataFrame. A polygon dataset for clipping the waveform}

\item{level2bdt}{level2bdt; S4 object of class "gedi.level1b.dt"}
}
\value{
Returns An object of class "gedi.level1b.dt"; subset of GEDI Level1B data
}
\description{
Clip GEDI level2bdt data within a given geometry area
}
\examples{

#' GEDI level1B file path
level1Bfilepath = system.file("extdata", "lvis_level1_clip.h5", package="rGEDI")

#' Reading GEDI level1B file
level1b = readLevel1b(level1Bfilepath)

#' Creating GEDI level2bdt object
level2bdt = level2bdt(level1b)

# Polgons file path
polygon_filepath <- system.file("extdata", "clip_polygon.shp", package="rGEDI")

# Reading GEDI level2B file
polygon_spdf<-raster::shapefile(polygons_filepath)

clipped_level2bdt = clipLevel1Geometry(level2bdt, polygon_spdf)

library(leaflet)
leaflet() \%>\%
 addCircleMarkers(clipped_level2bdt@dt$longitude_bin0,
                  clipped_level2bdt@dt$latitude_bin0,
                  radius = 1,
                  opacity = 1,
                  color = "red")  \%>\%
 addScaleBar(options = list(imperial = FALSE)) \%>\%
 addPolygons(data=polygon_spdf,weight=1,col = 'white',
             opacity = 1, fillOpacity = 0) \%>\%
 addProviderTiles(providers$Esri.WorldImagery)
}
>>>>>>> caio
