<<<<<<< HEAD
% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clipLevel2BPAVDProfile.R
\name{clipLevel2BPAVDProfile}
\alias{clipLevel2BPAVDProfile}
\title{Clip level2BPAVDProfile data by Coordinates}
\usage{
clipLevel2BPAVDProfile(level2BPAVDProfile, xleft, xright, ybottom, ytop, output="")
}
\arguments{
\item{xleft}{Numeric. West longitude (x) coordinate of bounding rectangle, in decimal degrees.}

\item{xright}{Numeric. East longitude (x) coordinate of bounding rectangle, in decimal degrees.}

\item{ybottom}{Numeric. South latitude (y) coordinate of bounding rectangle, in decimal degrees.}

\item{level2BPAVDProfile}{A GEDI Level2B object (output of \code{\link[rGEDI:getLevel2BPAVDProfile]{getLevel2BPAVDProfile}} function). A S4 object of class "gedi.level2b".}

\item{ytopNumeric.}{North latitude (y) coordinate of bounding rectangle, in decimal degrees.}

\item{output}{Optional character path where to save the new hdf5file. The default stores a temporary file only.}
}
\value{
An S4 object of class "gedi.level2b".
}
\description{
This function clips GEDI level2B-derived
Plant Area Volume Density profile within given bounding coordinates
}
\examples{
# specify the path and data file and read it
level2bpath <- system.file("extdata", "GEDIexample_level02B.h5", package="rGEDI")

# Reading GEDI level2B data
level2b <- readLevel2B(level2bpath)

# Get Plant Area Volume Density profile
level2BPAVDProfile<-getLevel2BPAVDProfile(level2b)

# Bounding rectangle coordinates
xleft = -116.4683
xright = -116.5583
ybottom = 46.75208
ytop = 46.84229

# clip level2BVPM by extent boundary box
level2BPAVDProfile_clip <- clipLevel2BPAVDProfile(level2BPAVDProfile,xleft, xright, ybottom, ytop)

}
=======
% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clipLevel2BPAVDProfile.R
\name{clipLevel2BPAVDProfile}
\alias{clipLevel2BPAVDProfile}
\title{Clip GEDI level2bdt data}
\usage{
clipLevel2BPAVDProfile(x, xleft, xright, ybottom, ytop)
}
\arguments{
\item{level2bdt}{level2bdt; S4 object of class "gedi.level1b.dt"}

\item{extent}{Extent object of a Spatial object .}
}
\value{
Returns An object of class "gedi.level1b.dt"; subset of GEDI Level1B data
}
\description{
Clip GEDI level2bdt data within a given bounding coordinates
}
\examples{

#' GEDI level1B file path
level1_filepath = system.file("extdata", "lvis_level1_clip.h5", package="rLVIS")

#' Reading LVIS level 2 file
level1_waveform = readLevel1b(level1Bfilepath)

# Rectangle area for cliping
xmin = -116.4683
xmax = -116.3177
ymin = 46.75208
ymax = 46.84229

# creating an exent object
ext<-extent(c(xmin=xmin, xmax=xmax, ymin=ymin, ymax=ymax))

clipped_level2bdt = clipLevel1(level2bdt, extent=ext)

library(leaflet)
leaflet() \%>\%
 addCircleMarkers(clipped_level2bdt@dt$longitude_bin0,
                  clipped_level2bdt@dt$latitude_bin0,
                  radius = 1,
                  opacity = 1,
                  color = "red")  \%>\%
 addScaleBar(options = list(imperial = FALSE)) \%>\%
 addProviderTiles(providers$Esri.WorldImagery)
}
>>>>>>> caio
