% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clipLevel2BPAVDProfile.R
\name{clipLevel2BPAVDProfileGeometry}
\alias{clipLevel2BPAVDProfileGeometry}
\title{Clip GEDI level2bdt data by geometry}
\usage{
clipLevel2BPAVDProfileGeometry(x, polygon_spdf, split_by = NULL)
}
\arguments{
\item{polygon_spdf}{SpatialDataFrame. A polygon dataset for clipping the waveform}

\item{level2bdt}{level2bdt; S4 object of class "gedi.level1b.dt"}
}
\value{
Returns An object of class "gedi.level1b.dt"; subset of GEDI Level1B data
}
\description{
Clip GEDI level2bdt data within a given geometry area
}
\examples{

#' GEDI level1B file path
level1Bfilepath = system.file("extdata", "lvis_level1_clip.h5", package="rGEDI")

#' Reading GEDI level1B file
level1b = readLevel1b(level1Bfilepath)

#' Creating GEDI level2bdt object
level2bdt = level2bdt(level1b)

# Polgons file path
polygon_filepath <- system.file("extdata", "clip_polygon.shp", package="rGEDI")

# Reading GEDI level2B file
polygon_spdf<-raster::shapefile(polygons_filepath)

clipped_level2bdt = clipLevel1Geometry(level2bdt, polygon_spdf)

library(leaflet)
leaflet() \%>\%
 addCircleMarkers(clipped_level2bdt@dt$longitude_bin0,
                  clipped_level2bdt@dt$latitude_bin0,
                  radius = 1,
                  opacity = 1,
                  color = "red")  \%>\%
 addScaleBar(options = list(imperial = FALSE)) \%>\%
 addPolygons(data=polygon_spdf,weight=1,col = 'white',
             opacity = 1, fillOpacity = 0) \%>\%
 addProviderTiles(providers$Esri.WorldImagery)
}
