% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clipLevel2BVPM.R
\name{clipLevel2BVPM}
\alias{clipLevel2BVPM}
\title{Clip level2BVPM data by Coordinates}
\usage{
clipLevel2BVPM(level2BVPM, xmin, xmax, ymin, ymax)
}
\arguments{
\item{level2BVPM}{A GEDI Level2B object (output of \code{\link[rGEDI:readLevel1B]{readLevel1B}} function). A S4 object of class "data.table".}

\item{xmin}{Numeric. West longitude (x) coordinate of bounding rectangle, in decimal degrees.}

\item{xmax}{Numeric. East longitude (x) coordinate of bounding rectangle, in decimal degrees.}

\item{ymin}{Numeric. South latitude (y) coordinate of bounding rectangle, in decimal degrees.}

\item{ymax}{Numeric. North latitude (y) coordinate of bounding rectangle, in decimal degrees.}
}
\value{
An S4 object of class \code{\link[data.table:data.table]{data.table-class}}
containing the Canopy Cover and Vertical Profile metrics.
}
\description{
This function clips GEDI level2B-derived
Canopy Cover and Vertical Profile metrics within given bounding coordinates
}
\examples{
\dontrun{
# specify the path to download GEDI example dataset
outdir<-getwd()

# downloading GEDI example dataset (zip file)
download.file(
             paste0(
                    "https://github.com/carlos-alberto-silva/rGEDI/",
                    "releases/download/examples/examples.zip"
             ),
             destfile=paste0(outdir,"/examples.zip"))

# unzip the file
unzip(paste0(outdir,"\\\\examples.zip"))

# specify the path to GEDI level2B data
level2bpath = paste0(outdir,"\\\\GEDI02_B_2019108080338_O01964_T05337_02_001_01_sub.h5")

# Reading GEDI level1B file
level2b<-readLevel2b(gedilevel2b)

# Get canopy cover and vertical profile metrics
level2BVPM<-getlevel2BVPM(level2b)

# Bounding rectangle coordinates
xmin = -44.15036
xmax = -44.10066
ymin = -13.75831
ymax = -13.71244

# clip level2BVPM by extent boundary box
level2b_clip <- level2BVPM(level2BVPM,xmin, xmax, ymin, ymax)

library(leaflet)
leaflet() \%>\%
 addCircleMarkers(level2b_clip@dt$longitude_bin0,
                  level2b_clip@dt$latitude_bin0,
                  radius = 1,
                  opacity = 1,
                  color = "red")  \%>\%
 addScaleBar(options = list(imperial = FALSE)) \%>\%
 addPolygons(data=polygon_spdf,weight=1,col = 'white',
             opacity = 1, fillOpacity = 0) \%>\%
 addProviderTiles(providers$Esri.WorldImagery)
}
}
\seealso{
https://lpdaac.usgs.gov/products/gedi02_bv001/
}
