% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clipLevel3.R
\name{clipLevel3}
\alias{clipLevel3}
\title{Clip GEDI Full Waveform Geolocations by Coordinates}
\usage{
clipLevel3(level3, xmin, xmax, ymin, ymax)
}
\arguments{
\item{level3}{A \code{\linkS4class{gedi.level3}} resulting from \code{\link[=readLevel3]{readLevel3()}}.}

\item{xmin}{Numeric. West longitude (x) coordinate of the bounding rectangle, in decimal degrees.}

\item{xmax}{Numeric. East longitude (x) coordinate of the bounding rectangle, in decimal degrees.}

\item{ymin}{Numeric. South latitude (y) coordinate of the bounding rectangle, in decimal degrees.}

\item{ymax}{Numeric. North latitude (y) coordinate of the bounding rectangle, in decimal degrees.}
}
\value{
Returns the clipped S4 object of class \code{\linkS4class{gedi.level3}}.
}
\description{
This function clips GEDI level3 extracted geolocation (\code{\link[=readLevel3]{readLevel3()}})
data a within given bounding coordinates
}
\examples{
# Specifying the path to GEDI level3 data (zip file)
outdir <- tempdir()

level3path <- "C:/Users/caiohamamura/Downloads/GEDI03_rh100_mean_2019108_2020287_002_02.tif"
level3 <- readLevel3(level3path = level3path)
xmin <- -5888627
xmax <- -5878615
ymin <- -990000
ymax <- -976400

# Clipping GEDI Full Waveform Geolocations by boundary box extent
level3_crop <- crop(level3, xmin, xmax, ymin, ymax)

if (require(ggplot2)) {
  ggplot() +
    geom_spatraster(data = level3_crop)
}

}
\seealso{
\url{https://daac.ornl.gov/GEDI/guides/GEDI_L3_LandSurface_Metrics_V2.html}
}
