% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clipLevel3.R
\name{clipLevel3}
\alias{clipLevel3}
\title{Clip GEDI Full Waveform Geolocations by Coordinates}
\usage{
clipLevel3(level3, xmin, xmax, ymin, ymax)
}
\arguments{
\item{level3}{A \code{\linkS4class{gedi.level3}} resulting from \code{\link[=getLevel3]{getLevel3()}}.}

\item{xmin}{Numeric. West longitude (x) coordinate of the bounding rectangle, in decimal degrees.}

\item{xmax}{Numeric. East longitude (x) coordinate of the bounding rectangle, in decimal degrees.}

\item{ymin}{Numeric. South latitude (y) coordinate of the bounding rectangle, in decimal degrees.}

\item{ymax}{Numeric. North latitude (y) coordinate of the bounding rectangle, in decimal degrees.}
}
\value{
Returns the clipped S4 object of class \code{\linkS4class{gedi.level3}}.
}
\description{
This function clips GEDI level3 extracted geolocation (\code{\link[=getLevel3]{getLevel3()}})
data a within given bounding coordinates
}
\examples{
# Specifying the path to GEDI level3 data (zip file)
outdir <- tempdir()
level3_fp_zip <- system.file("extdata",
  "GEDI01_B_2019108080338_O01964_T05337_02_003_01_sub.zip",
  package = "rGEDI"
)

# Unzipping GEDI level3 data
level3path <- unzip(level3_fp_zip, exdir = outdir)

# Reading GEDI level3 data (h5 file)
level3 <- readLevel3(level3path = level3path)

# Bounding rectangle coordinates
xmin <- -44.15036
xmax <- -44.10066
ymin <- -13.75831
ymax <- -13.71244

# Clipping GEDI Full Waveform Geolocations by boundary box extent
level3_clip <- clipLevel3(level3, xmin, xmax, ymin, ymax)

hasLeaflet <- require(leaflet)

if (hasLeaflet) {
  leaflet() \%>\%
    addCircleMarkers(level3_clip$longitude_bin0,
      level3_clip$latitude_bin0,
      radius = 1,
      opacity = 1,
      color = "red"
    ) \%>\%
    addScaleBar(options = list(imperial = FALSE)) \%>\%
    addProviderTiles(providers$Esri.WorldImagery)
}


}
\seealso{
\url{https://lpdaac.usgs.gov/products/gedi01_bv002/}
}
