% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clipLevel3.R
\name{clipLevel3Geometry}
\alias{clipLevel3Geometry}
\title{Clip GEDI Full Waveform Geolocations by geometry}
\usage{
clipLevel3Geometry(level3, polygon, split_by = "id")
}
\arguments{
\item{level3}{A \code{\linkS4class{gedi.level3}} resulting from \code{\link[=readLevel3]{readLevel3()}} function.}

\item{polygon}{Polygon. An object of class \code{SpatVect},
which can be loaded as an ESRI shapefile using \link[terra:vect]{terra::vect} function in the \emph{terra} package.}

\item{split_by}{Polygon id. If defined, GEDI data will be clipped by each polygon using the polygon id from
table of attribute defined by the user.}
}
\value{
Returns a list of S4 object of class \code{\link[data.table:data.table-class]{data.table::data.table}} containing the
clipped GEDI level3 extracted geolocations.
}
\description{
This function clips level3 extracted geolocation (level3)
data within a given geometry
}
\examples{
# Specifying the path to GEDI level3 data (zip file)
outdir <- tempdir()

level3path <- "C:/Users/caiohamamura/Downloads/GEDI03_rh100_mean_2019108_2020287_002_02.tif"
level3 <- readLevel3(level3path = level3path)


# Specifying the path to shapefile
polygon_filepath <- system.file("extdata", "stands_cerrado.shp", package = "rGEDI")

# Reading shapefile as SpatialPolygonsDataFrame object
library(terra)
polygon <- terra::vect(polygon_filepath)

# Clipping GEDI Full Waveform Geolocations by Geometry
level3_clip <- clipLevel3Geometry(level3, polygon, split_by = "id")

# Calculate global means
lapply(level3_clip, terra::global, mean)

# Plot all in a grid
n <- length(level3_clip)
sqrt_n <- ceiling(sqrt(n))
oldpar <- par()
par(mfrow = c(sqrt_n, sqrt_n))

for (l in level3_clip) {
  plot(l)
}
par(oldpar)

}
\seealso{
\url{https://daac.ornl.gov/GEDI/guides/GEDI_L3_LandSurface_Metrics_V2.html}
}
