% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clipLevel4A.R
\name{clipLevel4A}
\alias{clipLevel4A}
\title{Clip GEDI Level 4A by Coordinates}
\usage{
clipLevel4A(level4A, xmin, xmax, ymin, ymax)
}
\arguments{
\item{level4A}{A GEDI Level4A object (output of \code{\link[=readLevel4A]{readLevel4A()}} function).
An S4 object of class "data.table".}

\item{xmin}{Numeric. West longitude (x) coordinate of the bounding rectangle, in decimal degrees.}

\item{xmax}{Numeric. East longitude (x) coordinate of the bounding rectangle, in decimal degrees.}

\item{ymin}{Numeric. South latitude (y) coordinate of the bounding rectangle, in decimal degrees.}

\item{ymax}{Numeric. North latitude (y) coordinate of the bounding rectangle, in decimal degrees.}
}
\value{
Returns an S4 object of class \link[data.table:data.table]{data.table::data.table}
containing the AGBD and standard error estimates.
}
\description{
This function clips GEDI Level 4A data within given bounding coordinates
}
\examples{
# Specifying the path to GEDI level4A data (zip file)
outdir <- tempdir()
level4A_fp_zip <- system.file("extdata",
  "GEDI04_A_2019108080338_O01964_T05337_02_001_01_sub.zip",
  package = "rGEDI"
)

# Unzipping GEDI leve42A data
level4Apath <- unzip(level4A_fp_zip, exdir = outdir)

# Reading GEDI level2B data (h5 file)
level4a <- readLevel4A(level4Apath = level4Apath)

# Extracting canopy cover and vertical profile metrics
level4AVPM <- getLevel4AVPM(level4a)

# Bounding rectangle coordinates
xmin <- -44.15036
xmax <- -44.10066
ymin <- -13.75831
ymax <- -13.71244

# Clipping level4a by extent boundary box
level4a_clip <- clipLevel4a(level4a, xmin, xmax, ymin, ymax)

hasLeaflet <- require(leaflet)

if (hasLeaflet) {
  leaflet() \%>\%
    addCircleMarkers(level4a_clip$lon_lowestmode,
      level4a_clip$lat_lowestmode,
      radius = 1,
      opacity = 1,
      color = "red"
    ) \%>\%
    addScaleBar(options = list(imperial = FALSE)) \%>\%
    addProviderTiles(providers$Esri.WorldImagery)
}
close(level4a)
}
\seealso{
\url{https://daac.ornl.gov/GEDI/guides/GEDI_L4A_AGB_Density.html}
}
