% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clipLevel4A.R
\name{clipLevel4AGeometry}
\alias{clipLevel4AGeometry}
\title{Clip GEDI Level 4A by geometry}
\usage{
clipLevel4AGeometry(level4A, polygon_spdf, split_by)
}
\arguments{
\item{level4A}{A GEDI Level4A object (output of \code{\link[=readLevel4A]{readLevel4A()}} function).
An S4 object of class "gedi.level4a".}

\item{polygon_spdf}{Polygon. An object of class \code{\link[sp:SpatialPolygonsDataFrame-class]{sp::SpatialPolygonsDataFrame}},
which can be loaded as an ESRI shapefile using \link[raster:shapefile]{raster::shapefile} function in the \emph{raster} package.}

\item{split_by}{Polygon id. If defined, GEDI data will be clipped by each polygon using the attribute
specified by \code{split_by} from the attribute table.}
}
\value{
Returns an S4 object of class \link[data.table:data.table]{data.table::data.table}
containing the AGBD and standard error estiates.
}
\description{
This function clips GEDI Level 4A data within a given geometry
}
\examples{
# Specifying the path to GEDI level4A data (zip file)
outdir <- tempdir()
level4A_fp_zip <- system.file("extdata",
  "GEDI04_A_2019108080338_O01964_T05337_02_001_01_sub.zip",
  package = "rGEDI"
)

# Unzipping GEDI leve42A data
level4Apath <- unzip(level4A_fp_zip, exdir = outdir)

# Reading GEDI level2B data (h5 file)
level4a <- readLevel4A(level4Apath = level4Apath)

# Extracting canopy cover and vertical profile metrics
level4AVPM <- getLevel4AVPM(level4a)

# Specifying the path to shapefile
polygon_filepath <- system.file("extdata", "stands_cerrado.shp", package = "rGEDI")

# Reading shapefile as SpatialPolygonsDataFrame object
library(raster)
polygon_spdf <- shapefile(polygon_filepath)

# Clipping level4A by geometry
level4a_clip_geometry <- clipLevel4AGeometry(level4A, polygon_spdf, split_by = "id")

hasLeaflet <- require(leaflet)

if (hasLeaflet) {
  leaflet() \%>\%
    addCircleMarkers(level4a_clip$lon_lowestmode,
      level4a_clip$lat_lowestmode,
      radius = 1,
      opacity = 1,
      color = "red"
    ) \%>\%
    addScaleBar(options = list(imperial = FALSE)) \%>\%
    addProviderTiles(providers$Esri.WorldImagery)
}
close(level4a)
}
\seealso{
\url{https://daac.ornl.gov/GEDI/guides/GEDI_L4A_AGB_Density.html}
}
