% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clipLevel1BGeo.R
\name{clipxGeometry}
\alias{clipxGeometry}
\title{Clip GEDI x data by geometry}
\usage{
clipxGeometry(x, polygon_spdf, split_by = "id")
}
\arguments{
\item{x}{x; S4 object of class "gedi.level1b.dt"}

\item{polygon_spdf}{SpatialDataFrame. A polygon dataset for clipping the waveform}
}
\value{
Returns An object of class "gedi.level1b.dt"; subset of GEDI Level1B data
}
\description{
Clip GEDI x data within a given geometry area
}
\examples{

#' GEDI level1B file path
level1Bfilepath = system.file("extdata", "lvis_level1_clip.h5", package="rGEDI")

#' Reading GEDI level1B file
level1b = readLevel1b(level1Bfilepath)

#' Creating GEDI x object
x = x(level1b)

# Polgons file path
polygon_filepath <- system.file("extdata", "clip_polygon.shp", package="rGEDI")

# Reading GEDI level2B file
polygon_spdf<-raster::shapefile(polygons_filepath)

clipped_x = clipLevel1Geometry(x, polygon_spdf)

library(leaflet)
leaflet() \%>\%
 addCircleMarkers(clipped_x@dt$longitude_bin0,
                  clipped_x@dt$latitude_bin0,
                  radius = 1,
                  opacity = 1,
                  color = "red")  \%>\%
 addScaleBar(options = list(imperial = FALSE)) \%>\%
 addPolygons(data=polygon_spdf,weight=1,col = 'white',
             opacity = 1, fillOpacity = 0) \%>\%
 addProviderTiles(providers$Esri.WorldImagery)
}
