% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.gedi.R
\name{close}
\alias{close}
\alias{close,gedi.level1b-method}
\alias{close,gedi.level2a-method}
\alias{close,gedi.level2b-method}
\alias{close,gedi.level1bSim-method}
\title{Close hdf5 connections from gedi* objects}
\usage{
close(con, ...)

\S4method{close}{gedi.level1b}(con, ...)

\S4method{close}{gedi.level2a}(con, ...)

\S4method{close}{gedi.level2b}(con, ...)

\S4method{close}{gedi.level1bSim}(con, ...)
}
\arguments{
\item{con}{An object of class gedi*}

\item{...}{Inherited from base}
}
\description{
Close hdf5 connections from gedi* objects

Will close the gedi.level1b

Will close the gedi.level2a

Will close the gedi.level2b

Will close the gedi.level1bSim
}
