% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.gedi.R
\name{close}
\alias{close}
\alias{close,gedi.level1b-method}
\alias{close,gedi.level2a-method}
\alias{close,gedi.level2b-method}
\title{Close hdf5 connections from gedi* objects}
\usage{
close(con, ...)

\S4method{close}{gedi.level1b}(con, ...)

\S4method{close}{gedi.level2a}(con, ...)

\S4method{close}{gedi.level2b}(con, ...)
}
\arguments{
\item{con}{An object of class \verb{gedi.level*}}

\item{...}{Inherited from base}
}
\description{
Closing files will avoid locking HDF5 GEDI files.
}
