% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdalBindings.R
\name{formulaCalculate}
\alias{formulaCalculate}
\title{Calculate raster values based on a formula}
\usage{
formulaCalculate(formula, data, updateBand)
}
\arguments{
\item{formula}{Character. A textual formula to apply to the RasterBands}

\item{data}{List. A named list with the used variables in the textual formula}

\item{updateBand}{GDALRasterBand. The GDALRasterBand which will be updated with the calculated values.}
}
\value{
Nothing, it just updates the band of interest.
}
\description{
Calculate raster values based on a formula
}
