% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{GDALDataset}
\alias{GDALDataset}
\title{R6 Class GDALDataset wrapping}
\description{
Wrapping class for GDALDataset C++ API exporting GetRasterBand, GetRasterXSize, GetRasterYSize
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{GDALDataset$new()}}
\item \href{#method-GetRasterBand}{\code{GDALDataset$GetRasterBand()}}
\item \href{#method-GetRasterXSize}{\code{GDALDataset$GetRasterXSize()}}
\item \href{#method-GetRasterYSize}{\code{GDALDataset$GetRasterYSize()}}
\item \href{#method-clone}{\code{GDALDataset$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new raster file based on specified data. It will output a *.tif file.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GDALDataset$new(
  raster_path,
  ul_lat,
  ul_lon,
  lr_lat,
  lr_lon,
  res,
  datatype = GDALDataType$GDT_Float64,
  nbands = 1,
  projstring = defaultProjection,
  nodata = 0,
  co = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{raster_path}}{Character. The output path for the raster data}

\item{\code{ul_lat}}{Numeric. The upper left latitude.}

\item{\code{ul_lon}}{Numeric. The upper left longitude.}

\item{\code{lr_lat}}{Numeric. The lower right latitude.}

\item{\code{lr_lon}}{Numeric. The lower right longitude.}

\item{\code{res}}{Numeric. The resolution of the output raster}

\item{\code{datatype}}{GDALDataType. The GDALDataType to use for the raster, use (GDALDataType\$) to find the options. Default GDALDataType\$GDT_Float64}

\item{\code{nbands}}{Integer. Number of bands. Default 1.}

\item{\code{projstring}}{The projection string, either proj or WKT is accepted.}

\item{\code{nodata}}{Numeric. The no data value for the raster. Default 0.}

\item{\code{co}}{CharacterVector. A CharacterVector of creation options for GDAL. Default NULL}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An object from GDALDataset R6 class.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetRasterBand"></a>}}
\if{latex}{\out{\hypertarget{method-GetRasterBand}{}}}
\subsection{Method \code{GetRasterBand()}}{
Function to retrieve the GDALRasterBand R6 Object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GDALDataset$GetRasterBand(x)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{Integer. The band index, starting from 1 to number of bands.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An object of GDALRasterBand R6 class.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetRasterXSize"></a>}}
\if{latex}{\out{\hypertarget{method-GetRasterXSize}{}}}
\subsection{Method \code{GetRasterXSize()}}{
Get the width for the raster
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GDALDataset$GetRasterXSize()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
An integer indicating the raster width
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetRasterYSize"></a>}}
\if{latex}{\out{\hypertarget{method-GetRasterYSize}{}}}
\subsection{Method \code{GetRasterYSize()}}{
Get the height for the raster
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GDALDataset$GetRasterYSize()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
An integer indicating the raster height
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GDALDataset$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
