% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{GDALRasterBand}
\alias{GDALRasterBand}
\title{R6 Class GDALRasterBand wrapping
\loadmathjax}
\description{
Wrapping class for GDALRasterBand C++ API exporting ReadBlock, WriteBlock for better IO speed.
}
\note{
This constructor must not be called at all, this is automatically called from GDALDataset$GetRasterBand function.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{GDALRasterBand$new()}}
\item \href{#method-ReadBlock}{\code{GDALRasterBand$ReadBlock()}}
\item \href{#method-WriteBlock}{\code{GDALRasterBand$WriteBlock()}}
\item \href{#method-clone}{\code{GDALRasterBand$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Creates a new GDALRasterBand
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GDALRasterBand$new(band, datatype)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{band}}{The C++ pointer to the GDALRasterBandR object.}

\item{\code{datatype}}{The GDALDataType for this band}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An object of GDALRasterBand R6 class
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ReadBlock"></a>}}
\if{latex}{\out{\hypertarget{method-ReadBlock}{}}}
\subsection{Method \code{ReadBlock()}}{
Efficiently reads a raster block
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GDALRasterBand$ReadBlock(iXBlock, iYBlock)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{iXBlock}}{Integer. The i-th column block to access. The `iXBlock` will be offset \mjeqn{ BLOCKXSIZE \times iXBlock }{BLOCKXSIZE x iXBlock} from the origin.}

\item{\code{iYBlock}}{Integer. The i-th row block to access. The `iYBlock` will be offset \mjeqn{ BLOCKYSIZE \times iYBlock }{BLOCKYSIZE x iYBlock} from the origin.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
The returned Vector will be single dimensional with the length \mjeqn{ BLOCKXSIZE \times BLOCKYSIZE }{BLOCKXSIZE x BLOCKYSIZE}. If you use matrix(, ncol=BLOCKXSIZE) the matrix returned will actually be transposed. You should either transpose it or you can calculate the indices using \mjeqn{ y \cdot xsize + x }{y*xsize + x}
}

\subsection{Returns}{
RawVector for GDALDataType$GDT_Byte, IntegerVector for int types and NumericVector for floating point types.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-WriteBlock"></a>}}
\if{latex}{\out{\hypertarget{method-WriteBlock}{}}}
\subsection{Method \code{WriteBlock()}}{
Efficiently writes a raster block
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GDALRasterBand$WriteBlock(iXBlock, iYBlock, buffer)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{iXBlock}}{Integer. The i-th column block to write. The `iXBlock` will be offset \mjeqn{ BLOCKXSIZE \times iXBlock }{BLOCKXSIZE x iXBlock} from the origin.}

\item{\code{iYBlock}}{Integer. The i-th row block to write. The `iYBlock` will be offset \mjeqn{ BLOCKYSIZE \times iYBlock }{BLOCKYSIZE x iYBlock} from the origin.}

\item{\code{buffer}}{RawVector/IntegerVector/NumericVector depending on the GDALDataType. This should be a 1D vector with size equal to raster \mjeqn{ BLOCKXSIZE \times BLOCKYSIZE }{BLOCKXSIZE x BLOCKYSIZE}.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
The returned Vector will be single dimensional with the length \mjeqn{ BLOCKXSIZE \times BLOCKYSIZE }{BLOCKXSIZE x BLOCKYSIZE}. If you use matrix(, ncol=BLOCKXSIZE) the matrix returned will actually be transposed. You should either transpose it or you can calculate the indices using \mjeqn{ y \cdot xsize + x }{ y*xsize + x }.
}

\subsection{Returns}{
Nothing
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GDALRasterBand$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
