<<<<<<< HEAD:man/gedi.fullwaveform-class.Rd
% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.gedi.R
\docType{class}
\name{gedi.fullwaveform-class}
\alias{gedi.fullwaveform-class}
\alias{gedi.level2b}
\title{Class for GEDI level2B}
\description{
Class for GEDI level2B

Class for GEDI level1B Full Waveform
}
\section{Slots}{

\describe{
\item{\code{h5}}{Object of class H5File from hdf5r package}

\item{\code{dt}}{Object of class data.table from data.table package}
}}

=======
% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.gedi.R
\docType{class}
\name{gedi.fullwaveform-class}
\alias{gedi.fullwaveform-class}
\alias{gedi.level2b}
\title{Class for GEDI level2B}
\description{
Class for GEDI level2B

Class for GEDI level1B Full Waveform
}
\section{Slots}{

\describe{
\item{\code{h5}}{Object of class H5File from hdf5r package}

\item{\code{dt}}{Object of class data.table from data.table package}
}}

>>>>>>> caio:man/gedi.fullwaveform-class.Rd
