% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.gedi.level1b.R
\docType{class}
\name{gedi.level1b-class}
\alias{gedi.level1b-class}
\alias{gedi.level1b}
\title{Class for level1b from GEDI}
\description{
Class for level1b from GEDI
}
\section{Slots}{

\describe{
\item{\code{h5}}{Object of class H5File from hdf5r package}
}}

