% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.gedi.R
\docType{class}
\name{gedi.level1b-class}
\alias{gedi.level1b-class}
\alias{gedi.level1b}
\title{Class for GEDI level1B}
\description{
Class for GEDI level1B
}
\section{Slots}{

\describe{
\item{\code{h5}}{Object of class H5File from hdf5r package}
}}

