% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.gedi.R
\docType{class}
\name{gedi.level1b-class}
\alias{gedi.level1b-class}
\alias{gedi.level1b}
\title{Class for GEDI level1B}
\description{
Class for GEDI level1B
}
\section{Slots}{

\describe{
\item{\code{h5}}{Object of class \code{\link[hdf5r:H5File-class]{H5File}} from \code{hdf5r} package containing the
GEDI level1B products: geolocated Waveforms}
}}

\seealso{
\code{\link[hdf5r:H5File-class]{H5File}} in the \code{hdf5r} package and
\url{https://lpdaac.usgs.gov/products/gedi01_bv002/}
}
