% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.gedi.R
\docType{class}
\name{gedi.level2a-class}
\alias{gedi.level2a-class}
\alias{gedi.level2a}
\title{Class for GEDI level2A}
\description{
Class for GEDI level2A
}
\section{Slots}{

\describe{
\item{\code{h5}}{Object of class H5File from hdf5r package}
}}

