% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.gedi.R
\docType{class}
\name{gedi.level2a-class}
\alias{gedi.level2a-class}
\alias{gedi.level2a}
\title{Class for GEDI level2A}
\description{
Class for GEDI level2A
}
\section{Slots}{

\describe{
\item{\code{h5}}{Object of class H5File from \code{hdf5r} package containing the
GEDI level2A products: ground elevation, canopy top height, and relative heights (RH).}
}}

\seealso{
\code{\link[hdf5r:H5File-class]{H5File}} in the \code{hdf5r} package and
\url{https://lpdaac.usgs.gov/products/gedi02_av002/}
}
