% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.gedi.R
\docType{class}
\name{gedi.level2b-class}
\alias{gedi.level2b-class}
\alias{gedi.level2b}
\title{Class for GEDI level2B}
\description{
Class for GEDI level2B
}
\section{Slots}{

\describe{
\item{\code{h5}}{A local \code{\link[hdf5r:H5File-class]{H5File}} or cloud-backed GEDI HDF5
connection containing canopy cover, PAI, PAVD, and FHD.}
}}

\seealso{
\code{\link[hdf5r:H5File-class]{H5File}} in the \code{hdf5r} package and
\url{https://www.earthdata.nasa.gov/data/catalog/lpcloud-gedi02-b-002}
}
