% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.gedi.R
\docType{class}
\name{gedi.level2b-class}
\alias{gedi.level2b-class}
\alias{gedi.level2b}
\title{Class for GEDI level2B}
\description{
Class for GEDI level2B
}
\section{Slots}{

\describe{
\item{\code{h5}}{Object of class H5File from hdf5r package}
}}

