% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.gedi.R
\docType{class}
\name{gedi.level2b-class}
\alias{gedi.level2b-class}
\alias{gedi.level2b}
\title{Class for GEDI level2B}
\description{
Class for GEDI level2B
}
\section{Slots}{

\describe{
\item{\code{h5}}{Object of class H5File from \emph{hdf5r} package containing the
GEDI level2B products: canopy cover, Plant Area Index (PAI), Plant Area Volume Density (PAVD),
and Foliage Height Diversity (FHD).}
}}

\seealso{
\code{\link[hdf5r:H5File-class]{hdf5r::H5File}} in the \emph{hdf5r} package and
https://lpdaac.usgs.gov/products/gedi02_bv001/
}
