% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.gedi.R
\docType{class}
\name{gedi.level4a-class}
\alias{gedi.level4a-class}
\alias{gedi.level4a}
\title{Class for GEDI Level 4A footprint biomass data}
\description{
Class for GEDI Level 4A footprint biomass data
}
\section{Slots}{

\describe{
\item{\code{h5}}{An open \code{\link[hdf5r:H5File-class]{hdf5r::H5File}} containing a GEDI04_A granule.}
}}

\seealso{
\url{https://daac.ornl.gov/GEDI/guides/GEDI_L4A_AGB_Density_V3.html}
}
