% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.gedi.R
\docType{class}
\name{gedi.level4a-class}
\alias{gedi.level4a-class}
\alias{gedi.level4a}
\title{Class for GEDI level4A}
\description{
Class for GEDI level4A
}
\section{Slots}{

\describe{
\item{\code{h5}}{Object of class H5File from \code{hdf5r} package containing the
GEDI Level 4A (L4A): Predictions of the aboveground biomass density (AGBD; in Mg/ha) and
estimates of the prediction standard error within each sampled geolocated laser footprint.}
}}

\seealso{
\code{\link[hdf5r:H5File-class]{H5File}} in the \code{hdf5r} package and
\url{https://daac.ornl.gov/cgi-bin/dsviewer.pl?ds_id=1907}
}
