% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gedifinder.R
\name{GEDIfinder}
\alias{GEDIfinder}
\title{GEDI filder}
\usage{
GEDIfinder(x, xleft, xright, ybottom, ytop)
}
\arguments{
\item{x}{GEDI data type; "GEDI01_B", "GEDI02_A" or "GEDI02_B"}

\item{xleft}{numeric. left x coordinates of rectangles (degree).}

\item{xright}{numeric. right x coordinates of rectangles.}

\item{ybottom}{numeric. bottom y coordinates of rectangles.}

\item{ytop}{numeric. top y coordinates of rectangles.}
}
\value{
vector point out the path for download GEDI data within the bbox provided
}
\description{
This function finds the path to GEDI data within a boundary box provided
}
\examples{

gedi02_blist<-GEDIfinder(x="GEDI02_B",xleft=-40.9218,xright=-42.0248,ybottom=-71.7736,ytop=-73.7731)

}
\seealso{
gedifilder tool: https://lpdaacsvc.cr.usgs.gov/services/gedifinder
}
