% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gedifinder.R
\name{gedifinder}
\alias{gedifinder}
\title{GEDI finder}
\usage{
gedifinder(level, xleft, xright, ybottom, ytop)
}
\arguments{
\item{level}{GEDI data level; Options: "GEDI01_B", "GEDI02_A" or "GEDI02_B"}

\item{xleft}{numeric. left x coordinates of rectangles (degree).}

\item{xright}{numeric. right x coordinates of rectangles (degree).}

\item{ybottom}{numeric. bottom y coordinates of rectangles (degree).}

\item{ytop}{numeric. top y coordinates of rectangles (degree).}
}
\value{
vector object pointing out the path for download GEDI data within
the boundary box coordinates provided
}
\description{
This function finds the path to GEDI data within a boundary box coordinates provided

#'@usage gedifinder(level2BPAVDProfile, xleft, xright, ybottom, ytop, output="")
}
\examples{

# Bounding rectangle coordinates
xleft = -40.9218
xright = -42.0248
ybottom = -71.7736
ytop = -73.7731

# Getting the path to GEDI data within the provided boundary box coordinates
gedi02_b_list<-gedifinder(level="GEDI02_B",xleft=xleft,xright=xright,ybottom=ybottom,ytop=ytop)

}
\seealso{
This function just call the gedifilder tool developted by LPDAAC: https://lpdaacsvc.cr.usgs.gov/services/gedifinder
}
