% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gedifinder.R
\name{gedifinder}
\alias{gedifinder}
\title{GEDI finder}
\usage{
gedifinder(level, xmin, xmax, ymin, ymax)
}
\arguments{
\item{level}{GEDI data level; Options: "GEDI01_B", "GEDI02_A" or "GEDI02_B"}

\item{xmin}{Numeric. West longitude (x) coordinate of the bounding rectangle, in decimal degrees.}

\item{xmax}{Numeric. East longitude (x) coordinate of the bounding rectangle, in decimal degrees.}

\item{ymin}{Numeric. South latitude (y) coordinate of the bounding rectangle, in decimal degrees.}

\item{ymax}{Numeric. North latitude (y) coordinate of the bounding rectangle, in decimal degrees.}
}
\value{
Vector object pointing out the path saving the downloaded GEDI data within
the boundary box coordinates provided
}
\description{
This function finds the path to GEDI data within a boundary box coordinates provided

#'@usage gedifinder(level="GEDI02_B",xmin,xmax,ymin,ymax)
}
\examples{
\donttest{
# Specifying bounding box coordinates
xmin<- -44.17246
ymin<- -44.0654
xmax<- -13.76913
ymax<- -13.67646

# Extracting the path to GEDI data for the specified boundary box coordinates
gedi02b_list<-gedifinder(level="GEDI02_B",xmin,xmax,ymin,ymax)

}
}
\seealso{
This function just call the gedifinder tool developted by LPDAAC:
https://lpdaacsvc.cr.usgs.gov/services/gedifinder
}
