% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gedifinder.R
\name{gedifinder}
\alias{gedifinder}
\title{GEDI finder}
\usage{
gedifinder(
  product,
  ul_lat,
  ul_lon,
  lr_lat,
  lr_lon,
  version = "002",
  daterange = NULL
)
}
\arguments{
\item{product}{GEDI data level; Options: "GEDI01_B", "GEDI02_A" or "GEDI02_B"}

\item{ul_lat}{Numeric. Upper left (ul) corner coordinates, in lat
(decimal degrees) for the bounding box of the area of interest.}

\item{ul_lon}{Numeric. Upper left (ul) corner coordinates, in lon
(decimal degrees) for the bounding box of the area of interest.}

\item{lr_lat}{Numeric. Lower right (ul) corner coordinates, in lat
(decimal degrees) for the bounding box of the area of interest.}

\item{lr_lon}{Numeric. Lower right (ul) corner coordinates, in lon
(decimal degrees) for the bounding box of the area of interest.}

\item{version}{Character. The version of the GEDI product files to be
returned. Default "002".}

\item{daterange}{Vector. Date range. Specify your start and end dates
using ISO 8601 [YYYY]-[MM]-[DD]T[hh]:[mm]:[ss]Z. Ex.:
c("2019-07-01T00:00:00Z","2020-05-22T23:59:59Z"). If NULL (default),
the date range filter will be not applied.}
}
\value{
Return a vector object pointing out the path saving the downloaded
GEDI data within the boundary box coordinates provided
}
\description{
This function finds the exact granule(s) that contain GEDI data
for a given region of interest and date range
}
\examples{
\donttest{
# gedifinder is a web service provided by NASA
# usually the request takes more than 5 seconds

# Specifying bounding box coordinates
ul_lat <- 42.0
ul_lon <- -100
lr_lat <- 40.0
lr_lon <- -96.0

# Specifying the date range
daterange <- c("2019-07-01", "2020-05-22")

# Extracting the path to GEDI data for the specified boundary box coordinates
gedi02b_list <- gedifinder(
  product = "GEDI02_B",
  ul_lat,
  ul_lon,
  lr_lat,
  lr_lon,
  version = "002",
  daterange = daterange
)
}
}
\seealso{
bbox: Defined by the upper left and lower right corner coordinates,
in lat,lon ordering, for the bounding box of the area of interest
(e.g. [ul_lat,ul_lon,lr_lat,lr_lon]).

This function relies on the existing CMR tool:
\url{https://cmr.earthdata.nasa.gov/search/site/docs/search/api.html}
}
