% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C_simulateFullWaveform.R
\name{gediWFSimulator}
\alias{gediWFSimulator}
\title{GEDI full waveform data simulation}
\usage{
gediWFSimulator(input, output, waveID = NULL, coords = NULL,
  listCoord = NULL, gridBound = NULL, gridStep = 30, pSigma = -1,
  pFWHM = 15, readPulse = NULL, fSigma = 5.5, wavefront = NULL,
  res = 0.15, topHat = FALSE, sideLobe = FALSE, lobeAng = 0,
  checkCover = FALSE, maxScanAng = 1e+06, decimate = 1,
  pBuff = as.integer(2e+08), maxBins = as.integer(1024),
  countOnly = FALSE, pulseAfter = FALSE, pulseBefore = TRUE,
  noNorm = FALSE, noOctree = FALSE, octLevels = as.integer(0),
  nOctPix = as.integer(40), keepOld = FALSE, useShadow = FALSE,
  polyGround = FALSE, seed = NULL)
}
\arguments{
\item{input}{name. lasfile input filename}

\item{output}{name. output filename}

\item{waveID}{id. supply a waveID to pass to the output (only for single footprints)
Single footprint, list of footprints, or grid of footprints}

\item{listCoord}{name. Text file with list of coordinates. Pattern: X, Y, `[waveID]`, `[geoCoordsX]`, `[geoCoordsY]`. `[]` are optional.}

\item{gridStep}{res. grid step size
Lidar characteristics. Defaults are expected GEDI values. pSigmasig. set Gaussian pulse width as 1 sigma}

\item{pFWHM}{fhwm. set Gaussian pulse width as FWHM in ns}

\item{readPulse}{file. read pulse shape and width from a file insteda of making Gaussian}

\item{fSigma}{sig. set footprint width}

\item{wavefront}{file. read wavefront shape from file instead of setting Gaussian. Note that footprint width is still set by fSigma}

\item{res}{res. range resolution of waveform digitisation to output, in units of ALS data}

\item{topHat}{use a top hat wavefront}

\item{sideLobe}{use side lobes}

\item{lobeAng}{ang. lobe axis azimuth
Input data quality filters}

\item{checkCover}{check that the footprint is covered by ALS data. Do not output if not}

\item{maxScanAng}{ang. maximum scan angle, degrees}

\item{decimate}{x. probability of accepting an ALS beam
Computational speed options}

\item{pBuff}{s. point reading buffer size in Gbytes}

\item{maxBins}{for HDF5, limit number of bins to save trimming.}

\item{countOnly}{only use count method}

\item{pulseAfter}{apply the pulse smoothing after binning for computational speed, at the risk of aliasing (default)}

\item{pulseBefore}{apply the pulse smoothing before binning to avoid the risk of aliasing, at the expense of computational speed}

\item{noNorm}{don't normalise for ALS density
Octree}

\item{noOctree}{do not use an octree}

\item{octLevels}{n. number of octree levels to use}

\item{nOctPix}{n. number of octree pixels along a side for the top level
Using full-waveform input data (not tested)}

\item{keepOld}{do not overwrite old files, if they exist}

\item{useShadow}{account for shadowing in discrete return data through voxelisation}

\item{polyGround}{find mean ground elevation and slope through fitting a polynomial}

\item{seed}{n integer. random number seed}

\item{coord}{lon lat numeric vector. footprint coordinate in same system as lasfile}

\item{gridBounds}{minX maxX minY maxY numeric vector. make a grid of waveforms in this box}
}
\description{
Simulate GEDI full waveform data from Airborne Laser Scanning (ALS) 3-D point cloud

Input output filenames and format
}
