% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLevel2AM.R
\name{getLevel2AM}
\alias{getLevel2AM}
\title{Get GEDI Elevation and Height Metrics (GEDI Level2A)}
\usage{
getLevel2AM(level2a)
}
\arguments{
\item{level2a}{A GEDI Level2A object (output of \code{\link[rGEDI:readLevel2A]{readLevel2A}} function). A S4 object of class "gedi.level2a".}
}
\value{
An S4 object of class \code{\link[data.table:data.table]{data.table-class}}
containing the elevation and relative heights.
}
\description{
This function extracts Elevation and Relative Height (RH) metrics from GEDI Level2A data .
}
\examples{
\dontrun{
# specify the path to download GEDI example dataset
outdir<-getwd()

# downloading GEDI example dataset (zip file)
download.file(
             paste0(
                    "https://github.com/carlos-alberto-silva/rGEDI/",
                    "releases/download/examples/examples.zip"
             ),
             destfile=paste0(outdir,"/examples.zip"))

# unzip the file
unzip(paste0(outdir,"\\\\examples.zip"))

# specify the path to GEDI level2A data
level2apath = paste0(outdir,"\\\\GEDI02_A_2019108080338_O01964_T05337_02_001_01_sub.h5")

# Reading GEDI level2A data
level2a<-readLevel2A(level2apath)

# Get GEDI Elevation and Height Metrics
level2AM<-getLevel2AM(level2a)
head(level2AM)
}
}
\seealso{
https://lpdaac.usgs.gov/products/gedi02_av001/
}
