% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLevel2AM.R
\name{getLevel2AM}
\alias{getLevel2AM}
\title{Get Elevation and Height Metrics (GEDI Level2A)}
\usage{
getLevel2AM(level2a)
}
\arguments{
\item{level2a}{A GEDI Level2A object (output of \code{\link[rGEDI:readLevel2A]{readLevel2A}} function). A S4 object of class "gedi.level2a".}
}
\value{
An S4 object of class \code{\link[data.table:data.table]{data.table-class}}
containing the elevation and relative heights.
}
\description{
This function extracts Elevation and Relative Height (RH) metrics from GEDI Level2A data .
}
\examples{
# specify the path to GEDI Level 1B data
level2apath <- system.file("extdata", "GEDIexample_level01B.h5", package="rGEDI")

# Reading GEDI level1B data
level2a <- readLevel2A(level2apath)

# Get GEDI level1B geolocations
level2AM<-getLevel2AM(level2a)
head(level2AM)

}
