% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLevel2AM.R
\name{getLevel2AM}
\alias{getLevel2AM}
\title{Get GEDI Elevation and Height Metrics (GEDI Level2A)}
\usage{
getLevel2AM(level2a)
}
\arguments{
\item{level2a}{A GEDI Level2A object (output of \code{\link[rGEDI:readLevel2A]{readLevel2A}} function). A S4 object of class "gedi.level2a".}
}
\value{
An S4 object of class \code{\link[data.table:data.table]{data.table-class}}
containing the elevation and relative heights.
}
\description{
This function extracts Elevation and Relative Height (RH) metrics from GEDI Level2A data .
}
\examples{
\dontrun{
# specify the path to GEDI level2A data (zip file)
level2A_fp_zip <- system.file("extdata",
                  "GEDI02_A_2019108080338_O01964_T05337_02_001_01_sub.zip",
                  package="rGEDI")

# Unzipping GEDI level2A data
level2Apath <- unzip(level2A_fp_zip,exdir = dirname(level2A_fp_zip))

# Reading GEDI level2A data (h5 file)
level2a<-readLevel2A(level2Apath=level2Apath)

# Get GEDI Elevation and Height Metrics
level2AM<-getLevel2AM(level2a)
head(level2AM)
}
}
\seealso{
https://lpdaac.usgs.gov/products/gedi02_av001/
}
