% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLevel2BPAIProfile.R
\name{getLevel2BPAIProfile}
\alias{getLevel2BPAIProfile}
\title{Get Plant Area Index (PAI) Profile (GEDI Level2B)}
\usage{
getLevel2BPAIProfile(level2b)
}
\arguments{
\item{level2b}{A GEDI Level2B object (output of \code{\link[rGEDI:readLevel2B]{readLevel2B}} function). A S4 object of class "gedi.level2b".}
}
\value{
S4 object of class data.table;
}
\description{
This function extracts a Plant Area Index (PAI) Profile from GEDI Level2B data.
}
\seealso{
\code{\link[h5]{h5file}} in the \emph{h5} package.
}
