% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLevel2BPAIProfile.R
\name{getLevel2BPAIProfile}
\alias{getLevel2BPAIProfile}
\title{Read LVIS level2a data}
\usage{
getLevel2BPAIProfile(x)
}
\arguments{
\item{level2a}{H5File object from level2a}
}
\value{
H5File; S4 object of class H5File;
}
\description{
This function reads LVIS level2a data
}
\seealso{
\code{\link[h5]{h5file}} in the \emph{h5} package.
}
\author{
Carlos Alberto Silva. This function calls \emph{h5file} function from h5 package (Author: Mario Annau)
}
