% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLevel2BPAIProfile.R
\name{getLevel2BPAIProfile}
\alias{getLevel2BPAIProfile}
\title{Get GEDI Plant Area Index (PAI) Profile (GEDI Level2B)}
\usage{
getLevel2BPAIProfile(level2b)
}
\arguments{
\item{level2b}{A GEDI Level2B object (output of \code{\link[rGEDI:readLevel2B]{readLevel2B}} function). A S4 object of class "gedi.level2b".}
}
\value{
An S4 object of class \code{\link[data.table:data.table]{data.table-class}}
containing the elevation and relative heights.
}
\description{
This function extracts a Plant Area Index (PAI) Profile from GEDI Level2B data.
}
\examples{
\dontrun{
# specify the path to download GEDI example dataset
outdir<-getwd()

# downloading GEDI example dataset (zip file)
download.file(
             paste0(
                    "https://github.com/carlos-alberto-silva/rGEDI/",
                    "releases/download/examples/examples.zip"
             ),
             destfile=paste0(outdir,"/examples.zip"))

# unzip the file
unzip(paste0(outdir,"\\\\examples.zip"))

# specify the path to GEDI level2B data
level2bpath = paste0(outdir,"\\\\GEDI02_B_2019108080338_O01964_T05337_02_001_01_sub.h5")

# Reading GEDI level2B file
level2b<-readLevel2b(gedilevel2b)

# Get GEDI Plant Area Index (PAI) Profile (GEDI Level2B)
level2BPAIProfile<-getLevel2BPAIProfile(level2b)
head(level2BPAIProfile)
}
}
\seealso{
https://lpdaac.usgs.gov/products/gedi02_bv001/
}
