<<<<<<< HEAD
% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLevel2BPAVDProfile.R
\name{getLevel2BPAVDProfile}
\alias{getLevel2BPAVDProfile}
\title{Get Plant Area Volume Density (PAVD) Index (GEDI Level2B)}
\usage{
getLevel2BPAVDProfile(level2b)
}
\arguments{
\item{level2b}{A GEDI Level2B object (output of \code{\link[rGEDI:readLevel2B]{readLevel2B}} function). A S4 object of class "gedi.level2b".}
}
\value{
An S4 object of class \code{\link[data.table:data.table]{data.table-class}}
containing the PAVD Index.
}
\description{
This function extracts Plant Area Volume Density (PAVD) Index from GEDI Level2B data.
}
\examples{
# specify the path to GEDI Level 2B data
level2bpath <- system.file("extdata", "GEDIexample_level01B.h5", package="rGEDI")

# Reading GEDI level2B data
level2b <- readLevel2B(level2bpath)

# Get GEDI level2B geolocations
level2BPAVDProfile<-getLevel2BPAVDProfile(level2b)
head(level2BPAVDProfile)

}
=======
% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLevel2BPAVDProfile.R
\name{getLevel2BPAVDProfile}
\alias{getLevel2BPAVDProfile}
\title{Read LVIS level2a data}
\usage{
getLevel2BPAVDProfile(x)
}
\arguments{
\item{level2a}{H5File object from level2a}
}
\value{
H5File; S4 object of class H5File;
}
\description{
This function reads LVIS level2a data
}
\seealso{
\code{\link[h5]{h5file}} in the \emph{h5} package.
}
\author{
Carlos Alberto Silva. This function calls \emph{h5file} function from h5 package (Author: Mario Annau)
}
>>>>>>> caio
