% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLevel2BVPM.R
\name{getLevel2BVPM}
\alias{getLevel2BVPM}
\title{Get GEDI Canopy Cover and Vertical Profile Metrics (GEDI Level2B)}
\usage{
getLevel2BVPM(
  level2b,
  cols = c("beam", "shot_number", "algorithmrun_flag", "l2b_quality_flag", "delta_time",
    "sensitivity", "solar_elevation", "latitude_lastbin", "latitude_bin0",
    "longitude_bin0", "longitude_lastbin", "elev_highestreturn", "elev_lowestmode",
    "rh100", "pai", "fhd_normal", "omega", "pgap_theta", "cover")
)
}
\arguments{
\item{level2b}{A GEDI Level2B object (output of \code{\link[=readLevel2B]{readLevel2B()}}
function). An S4 object of class "gedi.level2b".}

\item{cols}{A character vector containing the list of columns to be extracted. See the default columns in the description.}
}
\value{
Returns an S4 object of class \link[data.table:data.table]{data.table::data.table}
containing the Vegetation Profile Biophysical Variables.
}
\description{
This function extracts information from GEDI Level2B data:
Total Plant Area Index,	Foliage Height Diversity, Foliage Clumping Index,
Total Gap Probability (theta), and Total canopy cover.
}
\details{
These are the biophysical variables and additional information extracted by default:
\itemize{
\item \emph{beam} Beam identifier
\item \emph{shot_number} Shot number
\item \emph{algorithmrun_flag} The L2B algorithm is run if this flag is set to 1 indicating data have sufficient waveform fidelity for L2B to run
\item \emph{l2b_quality_flag} L2B quality flag
\item \emph{delta_time} Transmit time of the shot since Jan 1 00:00 2018
\item \emph{sensitivity} Maxmimum canopy cover that can be penetrated
\item \emph{solar_elevation} Solar elevation
\item \emph{latitude_lastbin} Latitude of last bin of the pgap_theta_z, interpolated from L1B waveform coordinate
\item \emph{latitude_bin0} Latitude of first bin of the pgap_theta_z, interpolated from L1B waveform coordinate
\item \emph{elev_highestreturn} Elevation of highest detected return relative to reference ellipsoid
\item \emph{elev_lowestmode} Elevation of center of lowest mode relative to reference ellipsoid
\item \emph{rh100} RH100 slice
\item \emph{pai} Total Plant Area Index
\item \emph{fhd_normal} Foliage Height Diversity
\item \emph{omega} Foliage Clumping Index
\item \emph{pgap_theta} Total Gap Probability (theta)
\item \emph{cover} Total canopy cover
}

Every other columns in the GEDI2B product are also available, you can specify each column by using the \code{cols} parameter.
}
\examples{
# Specifying the path to GEDI level2B data (zip file)
outdir = tempdir()
level2B_fp_zip <- system.file("extdata",
                  "GEDI02_B_2019108080338_O01964_T05337_02_001_01_sub.zip",
                  package="rGEDI")

# Unzipping GEDI level2A data
level2Bpath <- unzip(level2B_fp_zip,exdir = outdir)

# Reading GEDI level2B data (h5 file)
level2b<-readLevel2B(level2Bpath=level2Bpath)

# Extracting GEDI Vegetation Profile Biophysical Variables
level2BVPM<-getLevel2BVPM(level2b, cols=c("beam", "shot_number"))
head(level2BVPM)

close(level2b)
}
\seealso{
\url{https://www.earthdata.nasa.gov/data/catalog/lpcloud-gedi02-b-002}
}
