% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLevel2BVPM.R
\name{getLevel2BVPM}
\alias{getLevel2BVPM}
\title{Get GEDI Vegetation Profile Biophysical Variables (GEDI Level2B)}
\usage{
getLevel2BVPM(level2b)
}
\arguments{
\item{level2b}{A GEDI Level2B object (output of \code{\link[rGEDI:readLevel2B]{readLevel2B}}
function). A S4 object of class "gedi.level2b".

#'@return An S4 object of class \code{\link[data.table:data.table]{data.table-class}}
containing the Vegetation Profile Biophysical Variables.}
}
\description{
This function extracts information from GEDI Level2B data:
Total Plant Area Index,	Foliage Height Diversity, Foliage Clumping Index,
Total Gap Probability (theta), and Total canopy cover.
}
\details{
These are the biophysical variables extracted:
\itemize{
\item \emph{pai} - Total Plant Area Index.
\item \emph{fhd_normal} -	Foliage Height Diversity.
\item \emph{omega} -	Foliage Clumping Index.
\item \emph{pgap_theta} -	Total Gap Probability (theta).
\item \emph{cover} -	Total canopy cover.
}
}
\examples{
# specify the path to GEDI Level 2B data
level2bpath <- system.file("extdata", "GEDIexample_level01B.h5", package="rGEDI")

# Reading GEDI level2B data
level2b <- readLevel2B(level2bpath)

# Get GEDI Vegetation Profile Biophysical Variables
level2BVPM<-getLevel2BVPM(level2b)
head(level2BVPM)

}
\seealso{
https://lpdaac.usgs.gov/products/gedi02_bv001/
}
