% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLevel2BVPM.R
\name{getLevel2BVPM}
\alias{getLevel2BVPM}
\title{Get GEDI Canopy Cover and Vertical Profile Metrics (GEDI Level2B)}
\usage{
getLevel2BVPM(level2b)
}
\arguments{
\item{level2b}{A GEDI Level2B object (output of \code{\link[rGEDI:readLevel2B]{readLevel2B}}
function). An S4 object of class "gedi.level2b".}
}
\value{
Returns an S4 object of class \code{\link[data.table:data.table]{data.table-class}}
containing the Vegetation Profile Biophysical Variables.
}
\description{
This function extracts information from GEDI Level2B data:
Total Plant Area Index,	Foliage Height Diversity, Foliage Clumping Index,
Total Gap Probability (theta), and Total canopy cover.
}
\details{
These are the biophysical variables extracted:
\itemize{
\item \emph{pai} Total Plant Area Index.
\item \emph{fhd_normal} Foliage Height Diversity.
\item \emph{omega} Foliage Clumping Index.
\item \emph{pgap_theta} Total Gap Probability (theta).
\item \emph{cover} Total canopy cover.
}
}
\examples{
# Specifying the path to GEDI level2B data (zip file)
outdir = tempdir()
level2B_fp_zip <- system.file("extdata",
                  "GEDI02_B_2019108080338_O01964_T05337_02_001_01_sub.zip",
                  package="rGEDI")

# Unzipping GEDI level2A data
level2Bpath <- unzip(level2B_fp_zip,exdir = outdir)

# Reading GEDI level2B data (h5 file)
level2b<-readLevel2B(level2Bpath=level2Bpath)

# Extracting GEDI Vegetation Profile Biophysical Variables
level2BVPM<-getLevel2BVPM(level2b)
head(level2BVPM)

close(level2b)
}
\seealso{
https://lpdaac.usgs.gov/products/gedi02_bv001/
}
