% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLevel4A.R
\name{getLevel4A}
\alias{getLevel4A}
\title{Get GEDI-derived Aboveground biomass density (GEDI Level4A)}
\usage{
getLevel4A(level4a)
}
\arguments{
\item{level4a}{A GEDI Level4A object (output of \code{\link[=readLevel4A]{readLevel4A()}} function).
An S4 object of class "gedi.level4a".}
}
\value{
Returns an S4 object of class \link[data.table:data.table]{data.table::data.table}
containing the aboveground biomass density (AGBD; in Mg/ha) and
standard error estimates.
}
\description{
This function extracts the aboveground biomass density (AGBD; in Mg/ha) and
standard error estimates within each sampled geolocated laser footprint.
}
\details{
Characteristics. Flag indicating likely invalid waveform (1=valid, 0=invalid).
\itemize{
\item \emph{beam} Beam identifier
\item \emph{shot_number} Shot number
\item \emph{algorithm_run_flag} The L4A algorithm is run if this flag is set to 1.
This flag selects data that have sufficient waveform fidelity for AGBD estimation.
\item \emph{l4_quality_flag} Flag simplifying selection of most useful biomass predictions
\item \emph{delta_time} Transmit time of the shot since Jan 1 00:00 2018
\item \emph{sensitivity} Maxmimum canopy cover that can be penetrated
\item \emph{solar_elevation} Solar elevation
\item \emph{lat_lowestmode} Latitude of center of lowest mode
\item \emph{lon_lowestmode} Longitude of center of lowest mode
\item \emph{elev_lowestmode} Elevation of center of lowest mode relative to reference ellipsoid
\item \emph{agbd} Aboveground biomass density (Mg/ha)
\item \emph{agbd_pi_upper} Lower prediction interval (see alpha attribute for the level)
\item \emph{agbd_pi_lower} Upper prediction interval (see alpha attribute for the level)
\item \emph{agbd_se} Aboveground biomass density (Mg/ha) prediction standard error
}
}
\examples{

# Specifying the path to GEDI level4A data (zip file)
outdir <- tempdir()
level4A_fp_zip <- system.file("extdata",
  "GEDI04_A_2019108080338_O01964_T05337_02_001_01_sub.zip",
  package = "rGEDI"
)

# Unzipping GEDI level4A data
level4Apath <- unzip(level4A_fp_zip, exdir = outdir)

# Reading GEDI level4A data (h5 file)
level4a <- readLevel2A(level4Apath = level4Apath)

# Extracting GEDI AGBD and standard error estimates.
level4AM <- getLevel4A(level4a)
head(level4AM)

close(level4a)
}
\seealso{
\url{https://daac.ornl.gov/cgi-bin/dsviewer.pl?ds_id=1907}
}
