% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridStatsLevel2BVPM.R
\name{gridStatsLevel2BVPM}
\alias{gridStatsLevel2BVPM}
\title{Compute Grids with descriptive statistics of
GEDI-derived Canopy Cover and Vertical Profile Metrics (Level2B)}
\usage{
gridStatsLevel2BVPM(level2VPM, func, res)
}
\arguments{
\item{level2BVPM}{A GEDI Level2AM object (output of \code{\link[rGEDI:getLevel2BVPM]{getLevel2BVPM}} function). A S4 object of class "data.table".}

\item{func}{the function(s) to be applied to each cell}

\item{res}{spatial resolution in decimal degrees for the output raster layer}
}
\value{
Returns raster layer(s) of selected GEDI Canopy Cover and Vertical Profile Metric(s)
}
\description{
This function computes a series of user-defined descriptive statistics within
each grid cell for GEDI-derived Canopy Cover and Vertical Profile Metrics (Level2B)
}
\examples{
\dontrun{
# specify the path to GEDI Level 2B data
level2bpath <- system.file("extdata", "GEDIexample_level02B.h5", package="rGEDI")

# Reading GEDI level2B data
level2b <- readLevel2B(level2bpath)

# Get GEDI-derived Canopy Cover and Vertical Profile Metrics
level2BVPM<-getLevel2BVPM(level2b)
head(level2BVPM)

#' Define your own function
mySetOfMetrics = function(x)
{
metrics = list(
   min =min(x), # Min of z
   max = max(x), # Max of z
   mean = mean(x), # Mean of z
   sd = sd(x)# Sd of z
   )
   return(metrics)
}

#' Computing a serie of statistics of GEDI-derived canopy cover
cover_stats<-gridStatsLevel2BVPM(level2AM = level2AM, func=mySetOfMetrics(cover), res=0.5)
plot(cover_stats)

#' Computing the max of the Total Plant Area Index only
pai_max<-gridStatsLevel2BVPM(level2AM = level2AM, func=max(pai), res=0.5)
plot(pai_max)

#' Computing the mean of Foliage Clumping Index only
omega_mean<-gridStatsLevel2BVPM(level2AM = level2AM, func=mean(omega), res=0.5)
plot(omega_mean)
}

}
