% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/level2BRasterize.R
\name{level2bRasterizeStats}
\alias{level2bRasterizeStats}
\title{Aggregate selected metrics into raster tif cells}
\usage{
level2bRasterizeStats(
  l2bDir,
  metrics,
  out_root,
  ul_lat,
  ul_lon,
  lr_lat,
  lr_lon,
  res,
  creation_options = c("COMPRESS=PACKBITS", "BIGTIFF=IF_SAFER", "TILED=YES",
    "BLOCKXSIZE=512", "BLOCKYSIZE=512")
)
}
\arguments{
\item{l2bDir}{CharacterVector. The directory paths where the H5 GEDI files are stored;}

\item{metrics}{CharacterVector. A vector of metrics available from Level2B, as in the getLevel2BVPM documentation}

\item{out_root}{Character. The root name for the raster output files, the pattern is {out_root}_{metric}_{count/m1/m2/m3/m4}.tif. This should include the full path for the file.}

\item{ul_lat}{Numeric. Upper left latitude for the bounding box}

\item{ul_lon}{Numeric. Upper left longitude for the bounding box}

\item{lr_lat}{Numeric. Lower right latitude for the bounding box}

\item{lr_lon}{Numeric. Lower right longitude for the bounding box}

\item{res}{Numeric. Resolution for the output raster in coordinates decimal degrees}

\item{creation_options}{CharacterVector. The GDAL creation options for the tif file. Default c("COMPRESS=PACKBITS", "BIGTIFF=IF_SAFER", "TILED=YES", "BLOCKXSIZE=512", "BLOCKYSIZE=512") will create BIGTIFF if needed, with PACKBITS compression and tiled by 512x512 pixels.}
}
\value{
Nothing. It outputs multiple raster tif files to the out_root specified path.
}
\description{
This function will read multiple Level2B H5 files and aggregate into multiple rasters: count, and 1st, 2nd, 3rd and 4th moments (count, m1, m2, m3 and m4) for each metric selected, from which we can calculate statistics such as Mean, SD, Skewness and Kurtosis.
}
