% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.gedi.R
\docType{methods}
\name{plot,gedi.fullwaveform,missing-method}
\alias{plot,gedi.fullwaveform,missing-method}
\title{Plot GEDI full-waveform}
\usage{
\S4method{plot}{gedi.fullwaveform,missing}(x, relative = FALSE,
  polygon = FALSE, ...)
}
\arguments{
\item{x}{An object of class "gedi.fullwaveform". (output of \code{\link[rGEDI:getLevel1BWF]{getLevel1BWF}} function)}

\item{relative}{if TRUE, the Wavform Amplitude will be showed in percentage (\%)}

\item{polygon}{if TRUE, polygon will be added to the plot}

\item{...}{will be passed to the main plot}
}
\description{
Plots a single GEDI full-waveform (level1b)
}
\examples{
\dontrun{
# specify the path to download GEDI example dataset
outdir<-getwd()

# downloading GEDI example dataset (zip file)
download.file(
             paste0(
                    "https://github.com/carlos-alberto-silva/rGEDI/",
                    "releases/download/examples/examples.zip"
             ),
             destfile=paste0(outdir,"/examples.zip"))

# unzip the file
unzip(paste0(outdir,"\\\\examples.zip"))

# specify the path to GEDI level1B data
level1bpath = paste0(outdir,"\\\\GEDI01_B_2019108080338_O01964_T05337_02_003_01_sub.h5")

# Reading GEDI level1B file
level1b<-readLevel1b(gedilevel1b)

# extract the desired information into a dataframe
wf <- getLevel1BWF(level1b, shot_number="19850022900500000")

# Plot Full-waveform
par(mfrow = c(1,2), cex.axis = 1.5)
plot(wf, relative=FALSE, polygon=TRUE, type="l", lwd=2, col="forestgreen",
xlab="", ylab="Elevation (m)")

plot(wf, relative=TRUE, polygon=TRUE, type="l", lwd=2, col="forestgreen",
xlab="Waveform Amplitude (\\\%)", ylab="Elevation (m)")
}
}
