% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.gedi.R
\docType{methods}
\name{plot,gedi.fullwaveform,missing-method}
\alias{plot,gedi.fullwaveform,missing-method}
\title{Plot GEDI full-waveform}
\usage{
plot(x, relative, polygon,...)
}
\arguments{
\item{x}{An object of class "gedi.fullwaveform". (output of \code{\link[rGEDI:getLevel1BWF]{getLevel1BWF}} function)}

\item{relative}{if TRUE, the Wavform Amplitude will be showed in percentage (%)}

\item{polygon}{if TRUE, polygon will be added to the plot}
}
\description{
Plots a single GEDI full-waveform (level1b)
}
