% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPAVDProfile.R
\name{plotPAVDProfile}
\alias{plotPAVDProfile}
\title{Plot Plant Area Volume Density Profile}
\usage{
plotPAVDProfile(level2BPAVDProfile, beam = "BEAM0101", elev = TRUE)
}
\arguments{
\item{level2BPAVDProfile}{A GEDI Level2B object (output of \code{\link[rGEDI:getLevel2BPAVDProfile]{getLevel2BPAVDProfile}} function). A S4 object of class "data.table".}

\item{beam}{select GEDI beam. Default is "BEAM0101". See details section.}

\item{elev}{if TRUE, elevation will be used for plotting the PAVD profile. Otherwise,
height will be used instead.}
}
\value{
Returns a ggplot object. See \code{\link[ggplot2:ggplot]{ggplot}} package.
}
\description{
This functions plots Plant Area Volume Density profile (GEDI level2B)
}
\details{
list of GEDI beams. See output of \code{\link[rGEDI:getLevel2BPAVDProfile]{getLevel2BPAVDProfile}} function.
\itemize{
\item \emph{BEAM0000}
\item \emph{BEAM0001}
\item \emph{BEAM0010}
\item \emph{BEAM0011}
\item \emph{BEAM0101}
\item \emph{BEAM0110}
\item \emph{BEAM1000}
\item \emph{BEAM1011}
}
}
\examples{
# specify the path to GEDI level2B data (zip file)
level2B_fp_zip <- system.file("extdata",
                  "GEDI02_B_2019108080338_O01964_T05337_02_001_01_sub.zip",
                  package="rGEDI")

# Unzipping GEDI level2A data
level2Bpath <- unzip(level2B_fp_zip,exdir = dirname(level2B_fp_zip))

# Reading GEDI level2B data (h5 file)
level2b<-readLevel2B(level2Bpath=level2Bpath)

# Get Plant Area Volume Density profile
level2BPAVDProfile<-getLevel2BPAVDProfile(level2b)

# Plot Level2B PAVD Profile
gprofile<-plotPAVDProfile(level2BPAVDProfile, beam="BEAM0101", elev=TRUE)



}
\seealso{
https://lpdaac.usgs.gov/products/gedi02_bv001/
}
