% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPAVDProfile.R
\name{plotPAVDProfile}
\alias{plotPAVDProfile}
\title{Plot Plant Area Volume Density Profile}
\usage{
plotPAVDProfile(level2BPAVDProfile, beam = "BEAM0101", elev = TRUE)
}
\arguments{
\item{level2BPAVDProfile}{A GEDI Level2B object (output of \code{\link[rGEDI:getLevel2BPAVDProfile]{getLevel2BPAVDProfile}} function). A S4 object of class "data.table".}

\item{beam}{select GEDI beam. Default is "BEAM0101". See details section.}

\item{elev}{if TRUE, elevation will be used for plotting the PAVD profile. Otherwise,
height will be used instead.}
}
\value{
Returns a ggplot object. See \code{\link[ggplot2:ggplot]{ggplot}}
}
\description{
This functions plots Plant Area Volume Density profile (GEDI level2B)
}
\details{
list of GEDI beams. See output of \code{\link[rGEDI:getLevel2BPAVDProfile]{getLevel2BPAVDProfile}} function.
\itemize{
\item \emph{BEAM0000}
\item \emph{BEAM0001}
\item \emph{BEAM0010}
\item \emph{BEAM0011}
\item \emph{BEAM0101}
\item \emph{BEAM0110}
\item \emph{BEAM1000}
\item \emph{BEAM1011}
}
}
\examples{
\dontrun{
# specify the path to download GEDI example dataset
outdir<-getwd()

# downloading GEDI example dataset (zip file)
download.file(
             paste0(
                    "https://github.com/carlos-alberto-silva/rGEDI/",
                    "releases/download/examples/examples.zip"
             ),
             destfile=paste0(outdir,"/examples.zip"))

# unzip the file
unzip(paste0(outdir,"\\\\examples.zip"))

# specify the path to GEDI level2B data
level2bpath = paste0(outdir,"\\\\GEDI02_B_2019108080338_O01964_T05337_02_001_01_sub.h5")

# Reading GEDI level1B file
level2b<-readLevel2b(gedilevel2b)

# Get Plant Area Volume Density profile
level2BPAVDProfile<-getLevel2BPAVDProfile(level2b)

# Plot Level2B PAVD Profile
gprofile<-plotPAVDProfile(level2BPAVDProfile, beam="BEAM0101", elev=TRUE)


}
}
\seealso{
https://lpdaac.usgs.gov/products/gedi02_bv001/
}
