% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotWFMetrics.R
\name{plotWFMetrics}
\alias{plotWFMetrics}
\title{GEDI-waveform plot with metrics}
\usage{
plotWFMetrics(level1b, level2a, shot_number, rh=c(25, 50, 75),...)
}
\arguments{
\item{level1b}{A GEDI Level1B object (output of \code{\link[rGEDI:readLevel1B]{readLevel1B}} function). A S4 object of class "gedi.level1b".}

\item{level2a}{A GEDI Level2A object (output of \code{\link[rGEDI:readLevel2A]{readLevel2A}} function). A S4 object of class "gedi.level2a".}

\item{shot_number}{Shot number. A scalar represeting the shot number of a giving pulse.}

\item{rh}{integer vector. Specify which RH metrics to plot except rh0 and rh100, default c(25, 50, 75).}

\item{...}{will be passed to the main plot.}
}
\value{
Returns raster layer(s) of selected GEDI Canopy Cover and Vertical Profile Metric(s)
}
\description{
Plots the waveform with overlaid RH metrics
}
\examples{
\dontrun{
# specify the path to download GEDI example dataset
outdir<-getwd()

# downloading GEDI example dataset (zip file)
download.file(
             paste0(
                    "https://github.com/carlos-alberto-silva/rGEDI/",
                    "releases/download/examples/examples.zip"
             ),
             destfile=paste0(outdir,"/examples.zip"))

# unzip the file
unzip(paste0(outdir,"\\\\examples.zip"))

# specify the path to GEDI level1B and level2A data
level1bpath = paste0(outdir,"\\\\GEDI01_B_2019108080338_O01964_T05337_02_003_01_sub.h5"))
level2apath = paste0(outdir,"\\\\GEDI02_A_2019108080338_O01964_T05337_02_001_01_sub.h5"))

# Reading GEDI level1B and level2A files
level1b<-readLevel1b(gedilevel1b)
level2a<-readLevel2a(gedilevel2a)

shot_number = "19640521100108408"

plotWFMetrics(level1b, level2a, shot_number, rh=c(25, 50, 75, 90))
}

}
\seealso{
https://lpdaac.usgs.gov/products/gedi02_bv001/
}
