% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readLevel1b.R
\name{readLevel1B}
\alias{readLevel1B}
\title{Read GEDI Level1B data}
\usage{
readLevel1B(level1Bpath)
}
\arguments{
\item{level1Bpath}{file path pointing to GEDI level1B data (H5 format)}
}
\value{
level1BFile; S4 object of class "gedi.level1b";
}
\description{
This function reads GEDI level1B data
}
