% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readLevel1b.R
\name{readLevel1B}
\alias{readLevel1B}
\title{Read Geolocated Waveforms (GEDI Level1B)}
\usage{
readLevel1B(level1Bpath)
}
\arguments{
\item{level1Bpath}{file path pointing to GEDI level1B data. Data in HDF5 Hierarchical Data Format (.h5).}
}
\value{
S4 object of class "gedi.level1b".
}
\description{
This function reads GEDI level1B products: geolocated Waveforms
}
\examples{
# specify the path and data file to be read
level1bpath <- system.file("extdata", "GEDIexample_level01B.h5", package="rGEDI")

# read the file
gedilevel1b<-readLevel1B(level1bpath)


}
\seealso{
\code{\link[hdf5r]{hdf5rfile}} in the \emph{hdf5r} package and
https://lpdaac.usgs.gov/products/gedi01_bv001/
}
