% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readLevel1b.R
\name{readLevel1b}
\alias{readLevel1b}
\title{Read LVIS Level1b data}
\usage{
readLevel1b(level1bpath)
}
\arguments{
\item{level1bpath}{file path pointing to LVIS level1b data (H5 format)}
}
\value{
H5File; S4 object of class H5File;
}
\description{
This function reads LVIS level1b data
}
\seealso{
\code{\link[h5]{h5file}} in the \emph{h5} package.
}
\author{
Carlos Alberto Silva. This function calls \emph{h5file} function from h5 package (Author: Mario Annau)
}
