% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readLevel1B.R
\name{readLevel1B}
\alias{readLevel1B}
\title{Read Geolocated Waveforms (GEDI Level1B)}
\usage{
readLevel1B(level1Bpath)
}
\arguments{
\item{level1Bpath}{file path pointing to GEDI level1B data. Data in HDF5 Hierarchical Data Format (.h5).}
}
\value{
S4 object of class "gedi.level1b".
}
\description{
This function reads GEDI level1B products: geolocated Waveforms
}
\examples{
\dontrun{
# specify the path to download GEDI example dataset
outdir<-getwd()

# downloading GEDI example dataset (zip file)
download.file(paste0(
                    "https://github.com/carlos-alberto-silva/rGEDI",
                    "/releases/download/examples/examples.zip"
                   ),
             destfile=outdir)

# unzip the file
unzip(paste0(outdir,"\\\\examples.zip"))

# specify the path to GEDI level1B data
level1bpath = paste0(outdir,"\\\\GEDI01_B_2019108080338_O01964_T05337_02_003_01_sub.h5")

# Reading GEDI level1B file
level1b<-readLevel1b(gedilevel1b)

}
}
\seealso{
\code{\link[hdf5r:H5File-class]{hdf5r::H5File}} in the \emph{hdf5r} package and
https://lpdaac.usgs.gov/products/gedi01_bv001/
}
