% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readLevel2A.R
\name{readLevel2A}
\alias{readLevel2A}
\title{Read GEDI Level2A data (Basic Full Waveform derived Metrics)}
\usage{
readLevel2A(level2Apath)
}
\arguments{
\item{level2Apath}{Local file path or Earthdata Cloud URL pointing to a
GEDI Level 2A HDF5 granule.}
}
\value{
Returns an S4 object of class \code{\link[=gedi.level2a-class]{gedi.level2a}} containing GEDI level2A data.
}
\description{
This function reads GEDI level2A products: ground elevation, canopy top height, and relative heights (RH).
}
\examples{
# Specifying the path to GEDI level2A data (zip file)
outdir = tempdir()
level2A_fp_zip <- system.file("extdata",
                  "GEDI02_A_2019108080338_O01964_T05337_02_001_01_sub.zip",
                  package="rGEDI")

# Unzipping GEDI level2A data
level2Apath <- unzip(level2A_fp_zip,exdir = outdir)

# Reading GEDI level2A data (h5 file)
level2a<-readLevel2A(level2Apath=level2Apath)

close(level2a)
}
\seealso{
\url{https://www.earthdata.nasa.gov/data/catalog/lpcloud-gedi02-a-002}
}
