% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readLevel2A.R
\name{readLevel2A}
\alias{readLevel2A}
\title{Read GEDI Level2A data}
\usage{
readLevel2A(level2Apath)
}
\arguments{
\item{level2bpath}{file path pointing to GEDI level2B data (H5 format)}
}
\value{
level2AFile; S4 object of class "gedi.level2a";
}
\description{
This function reads GEDI level2A data
}
