% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readLevel2B.R
\name{readLevel2B}
\alias{readLevel2B}
\title{Read GEDI Biophysical Metrics (Level2B data)}
\usage{
readLevel2B(level2Bpath)
}
\arguments{
\item{level2Bpath}{file path pointing to GEDI level2B data. Data in HDF5 Hierarchical Data Format (.h5).}
}
\value{
S4 object of class "gedi.level1b".
}
\description{
This function reads GEDI level2B products: canopy cover, Plant Area Index (PAI), Plant Area Volume Density (PAVD), and Foliage Height Diversity (FHD).

#'
}
\examples{
\dontrun{
# specify the path to download GEDI example dataset
outdir<-getwd()

# downloading GEDI example dataset (zip file)
download.file(paste0(
                    "https://github.com/carlos-alberto-silva/rGEDI",
                    "/releases/download/examples/examples.zip"
                   ),
             destfile=outdir)

# unzip the file
unzip(paste0(outdir,"\\\\examples.zip"))

# specify the path to GEDI level2B data
level2bpath = paste0(outdir,"\\\\GEDI02_B_2019108080338_O01964_T05337_02_001_01_sub.h5")

# Reading GEDI level1B file
level2b<-readLevel2b(gedilevel2b)

}
}
