% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readLevel2B.R
\name{readLevel2B}
\alias{readLevel2B}
\title{Read GEDI Level2B data}
\usage{
readLevel2B(level2Bpath)
}
\arguments{
\item{level2Bpath}{file path pointing to GEDI level2B data (H5 format)}
}
\value{
level2bFile; S4 object of class "gedi.level2b";
}
\description{
This function reads GEDI level2B data
}
