% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readLevel3.R
\name{readLevel3}
\alias{readLevel3}
\title{Read GEDI Level3 data (Rasterized metrics)}
\usage{
readLevel3(level3path)
}
\arguments{
\item{level3path}{File path pointing to GEDI level3 data. Data in HDF5 Hierarchical Data Format (.h5).}
}
\value{
Returns an S4 object of class \code{\linkS4class{gedi.level3}} containing GEDI level3 data.
}
\description{
This function reads GEDI level3 products: rasterized metrics
}
\examples{
# Specifying the path to GEDI level3 data (zip file)
outdir = tempdir()
level3_fp_zip <- system.file("extdata",
                  "GEDI01_B_2019108080338_O01964_T05337_02_003_01_sub.zip",
                  package="rGEDI")

# Unzipping GEDI level3 data
level3path <- unzip(level3_fp_zip,exdir = outdir)

# Reading GEDI level3 data (h5 file)
level3<-readLevel3(level3path=level3path)

close(level3)
}
\seealso{
\code{\link[hdf5r:H5File-class]{hdf5r::H5File}} in the \emph{hdf5r} package and
\url{https://lpdaac.usgs.gov/products/gedi01_bv002/}
}
