% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readLevel4A.R
\name{readLevel4A}
\alias{readLevel4A}
\title{Read GEDI Level4A data}
\usage{
readLevel4A(level4Apath)
}
\arguments{
\item{level2Apath}{File path pointing to GEDI level2A data. Data in HDF5 Hierarchical Data Format (.h5).}
}
\value{
Returns an S4 object of class \code{\linkS4class{gedi.level4a}} containing GEDI level4A data.
}
\description{
This function reads GEDI Level 4A (L4A) products: Predictions of the
aboveground biomass density (AGBD; in Mg/ha) and estimates of the prediction standard
error within each sampled geolocated laser footprint.
}
\examples{
# Specifying the path to GEDI level4A data (zip file)
outdir <- tempdir()
level4A_fp_zip <- system.file("extdata",
  "GEDI04_A_2019108080338_O01964_T05337_02_001_01_sub.zip",
  package = "rGEDI"
)

# Unzipping GEDI level4A data
level4Apath <- unzip(level4A_fp_zip, exdir = outdir)

# Reading GEDI level4A data (h5 file)
level2a <- readLevel4A(level4Apath = level4Apath)

close(level4a)
}
\seealso{
\url{https://daac.ornl.gov/cgi-bin/dsviewer.pl?ds_id=1907}
}
