% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rGEDI-package.R
\docType{package}
\name{rGEDI-package}
\alias{rGEDI-package}
\title{rGEDI: An R Package for NASA's Global Ecosystem Dynamics Investigation (GEDI) Data Visualizing and Processing.
See more details about GEDI data in https://gedi.umd.edu/data/products/.}
\description{
rGEDI: An R Package for NASA's Global Ecosystem Dynamics Investigation (GEDI) Data Visualizing and Processing.
See more details about GEDI data in https://gedi.umd.edu/data/products/.
}
\author{
Carlos A. Silva, Caio Hamamura, Ruben Valbuena, Steve Hancock, Adrian Cardil, Eben N. Broadbent, Danilo R. A. de Almeida, Celso H. L. Silva Junior and Carine Klauberg

The rGEDI package provides functions for i) downloading, ii) visualizing, iii) clipping, iv) gridding, iv) simulating and v) exporting GEDI data.

For comprehensive examples refer to \href{../doc/tutorial.html}{\code{vignette("Tutorial", package = "rGEDI")}}
}
